refactor(fbuild-daemon): split handlers/emulator.rs into submodules to satisfy LOC gate
