--- crates/fbuild-daemon/src/handlers/emulator.rs
@@ -1,3245 +1,0 @@
-use crate::context::DaemonContext;
-use crate::handlers::operations::{MonitorOutcome, MonitorState};
-use crate::models::OperationResponse;
-use axum::extract::{Path as AxumPath, State};
-use axum::http::{header, HeaderValue, StatusCode};
-use axum::response::{Html, IntoResponse};
-use axum::Json;
-use fbuild_packages::{Package, Toolchain};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::path::{Path, PathBuf};
-use std::process::Stdio;
-use std::sync::Arc;
-use tokio::io::{AsyncBufReadExt, BufReader};
-
-const AVR8JS_APP_JS: &str = include_str!("../../web/avr8js/app.js");
-const AVR8JS_HEADLESS_MJS: &str = include_str!("../../web/avr8js/headless.mjs");
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct Avr8jsSessionManifest {
-    session_id: String,
-    project_dir: String,
-    env_name: String,
-    board_id: String,
-    platform: String,
-    mcu: String,
-    f_cpu_hz: u32,
-    firmware_hex: String,
-    firmware_elf: Option<String>,
-    created_at_unix: f64,
-}
-
-#[derive(Debug, Serialize)]
-struct Avr8jsSessionResponse {
-    session_id: String,
-    env_name: String,
-    board_id: String,
-    platform: String,
-    mcu: String,
-    f_cpu_hz: u32,
-    firmware_hex_url: String,
-}
-
-fn now_unix() -> f64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs_f64()
-}
-
-fn render_page(session_id: &str) -> String {
-    format!(
-        r#"<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>AVR8js Monitor</title>
-  <style>
-    :root {{
-      color-scheme: light;
-      --bg: #f4efe4;
-      --panel: #fffdf8;
-      --ink: #1f1b18;
-      --muted: #6d6258;
-      --line: #d9cdbd;
-      --accent: #bd4b2c;
-    }}
-    body {{
-      margin: 0;
-      min-height: 100vh;
-      background:
-        radial-gradient(circle at top left, #fff7ea 0%, rgba(255,247,234,0) 35%),
-        linear-gradient(160deg, #f7f0e5 0%, #efe4d5 100%);
-      color: var(--ink);
-      font-family: "Consolas", "Menlo", monospace;
-      display: grid;
-      place-items: center;
-      padding: 24px;
-      box-sizing: border-box;
-    }}
-    .shell {{
-      width: min(920px, 100%);
-      background: var(--panel);
-      border: 1px solid var(--line);
-      box-shadow: 0 18px 60px rgba(74, 52, 34, 0.16);
-      border-radius: 18px;
-      overflow: hidden;
-    }}
-    .bar {{
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      gap: 16px;
-      padding: 14px 18px;
-      border-bottom: 1px solid var(--line);
-      background: linear-gradient(180deg, #fffaf2 0%, #f8efe2 100%);
-    }}
-    .title {{
-      font-size: 14px;
-      font-weight: 700;
-      letter-spacing: 0.08em;
-      text-transform: uppercase;
-    }}
-    .status {{
-      font-size: 12px;
-      color: var(--muted);
-    }}
-    .stdout {{
-      width: 100%;
-      min-height: 68vh;
-      border: 0;
-      outline: none;
-      resize: none;
-      padding: 18px;
-      box-sizing: border-box;
-      background: #231c16;
-      color: #f4efe8;
-      font: 15px/1.5 "Consolas", "Menlo", monospace;
-    }}
-    .stdout::selection {{
-      background: rgba(255, 190, 120, 0.35);
-    }}
-    .footer {{
-      display: flex;
-      justify-content: space-between;
-      gap: 16px;
-      padding: 10px 18px 16px;
-      border-top: 1px solid var(--line);
-      font-size: 12px;
-      color: var(--muted);
-      background: #fffaf2;
-    }}
-    .accent {{
-      color: var(--accent);
-      font-weight: 700;
-    }}
-  </style>
-</head>
-<body>
-  <div class="shell">
-    <div class="bar">
-      <div class="title">AVR8js Emulator</div>
-      <div class="status" id="status">Loading session...</div>
-    </div>
-    <textarea id="stdout" class="stdout" readonly spellcheck="false"></textarea>
-    <div class="footer">
-      <div>Session <span class="accent">{}</span></div>
-      <div>Browser-side ATmega328P serial monitor</div>
-    </div>
-  </div>
-  <script>window.__AVR8JS_SESSION_ID__ = {};</script>
-  <script type="module" src="/emulator/avr8js/app.js"></script>
-</body>
-</html>
-"#,
-        session_id,
-        serde_json::to_string(session_id).unwrap_or_else(|_| "\"\"".to_string()),
-    )
-}
-
-fn load_session_manifest(
-    ctx: &DaemonContext,
-    session_id: &str,
-) -> fbuild_core::Result<Avr8jsSessionManifest> {
-    let manifest_path = ctx
-        .avr8js_sessions
-        .get(session_id)
-        .map(|entry| entry.value().clone())
-        .ok_or_else(|| {
-            fbuild_core::FbuildError::Other(format!("unknown AVR8js session '{}'", session_id))
-        })?;
-    let raw = std::fs::read_to_string(&manifest_path)?;
-    serde_json::from_str(&raw).map_err(|e| {
-        fbuild_core::FbuildError::Other(format!("failed to parse AVR8js session manifest: {}", e))
-    })
-}
-
-pub struct DeployAvr8jsRequest {
-    pub request_id: String,
-    pub project_dir: PathBuf,
-    pub env_name: String,
-    pub board_id: String,
-    pub platform: fbuild_core::Platform,
-    pub firmware_path: PathBuf,
-    pub elf_path: Option<PathBuf>,
-    pub monitor_after: bool,
-    pub output_file: String,
-    pub output_dir: Option<String>,
-    pub monitor_timeout: Option<f64>,
-    pub halt_on_error: Option<String>,
-    pub halt_on_success: Option<String>,
-    pub expect: Option<String>,
-    pub show_timestamp: bool,
-    pub verbose: bool,
-}
-
-pub struct DeployQemuRequest {
-    pub request_id: String,
-    pub project_dir: PathBuf,
-    pub env_name: String,
-    pub board_id: String,
-    pub platform: fbuild_core::Platform,
-    pub firmware_path: PathBuf,
-    pub elf_path: Option<PathBuf>,
-    pub output_file: String,
-    pub output_dir: Option<String>,
-    pub monitor_timeout: Option<f64>,
-    pub qemu_timeout_secs: u32,
-    pub halt_on_error: Option<String>,
-    pub halt_on_success: Option<String>,
-    pub expect: Option<String>,
-    pub show_timestamp: bool,
-    pub verbose: bool,
-    pub board_overrides: HashMap<String, String>,
-}
-
-struct ProcessLine {
-    is_stderr: bool,
-    line: String,
-}
-
-enum ProcessEvent {
-    Line(ProcessLine),
-    StreamClosed,
-}
-
-struct QemuRunResult {
-    outcome: MonitorOutcome,
-    stdout: String,
-    stderr: String,
-    exit_code: Option<i32>,
-}
-
-struct RunQemuOptions<'a> {
-    elf_path: Option<PathBuf>,
-    addr2line_path: Option<PathBuf>,
-    timeout_secs: Option<f64>,
-    halt_on_error: Option<&'a str>,
-    halt_on_success: Option<&'a str>,
-    expect: Option<&'a str>,
-    show_timestamp: bool,
-    verbose: bool,
-    /// Label used in user-visible messages (e.g. "QEMU", "simavr").
-    process_label: &'a str,
-}
-
-pub async fn deploy_avr8js(
-    ctx: Arc<DaemonContext>,
-    req: DeployAvr8jsRequest,
-) -> (StatusCode, Json<OperationResponse>) {
-    let DeployAvr8jsRequest {
-        request_id,
-        project_dir,
-        env_name,
-        board_id,
-        platform,
-        firmware_path,
-        elf_path,
-        monitor_after,
-        output_file,
-        output_dir,
-        monitor_timeout,
-        halt_on_error,
-        halt_on_success,
-        expect,
-        show_timestamp,
-        verbose,
-    } = req;
-
-    if !matches!(
-        platform,
-        fbuild_core::Platform::AtmelAvr | fbuild_core::Platform::AtmelMegaAvr
-    ) {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                "avr8js deploy target is only supported for AVR boards".to_string(),
-            )),
-        );
-    }
-    if board_id != "uno" {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                format!(
-                    "avr8js deploy target currently supports only board 'uno' (got '{}')",
-                    board_id
-                ),
-            )),
-        );
-    }
-    if firmware_path.extension().and_then(|ext| ext.to_str()) != Some("hex") {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                format!(
-                    "avr8js deploy target requires firmware.hex, got '{}'",
-                    firmware_path.display()
-                ),
-            )),
-        );
-    }
-
-    let board = match fbuild_config::BoardConfig::from_board_id(&board_id, &HashMap::new()) {
-        Ok(board) => board,
-        Err(e) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("failed to load AVR8js board config: {}", e),
-                )),
-            );
-        }
-    };
-    if !board.mcu.eq_ignore_ascii_case("atmega328p") {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                format!(
-                    "avr8js deploy target currently supports only ATmega328P, got '{}'",
-                    board.mcu
-                ),
-            )),
-        );
-    }
-
-    let session_id = uuid::Uuid::new_v4().to_string();
-    let session_dir = fbuild_paths::get_project_fbuild_dir(&project_dir)
-        .join("emulators")
-        .join("avr8js")
-        .join(&env_name)
-        .join(&session_id);
-    if let Err(e) = std::fs::create_dir_all(&session_dir) {
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(OperationResponse::fail(
-                request_id,
-                format!("failed to create AVR8js session dir: {}", e),
-            )),
-        );
-    }
-
-    let staged_hex = session_dir.join("firmware.hex");
-    if let Err(e) = std::fs::copy(&firmware_path, &staged_hex) {
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(OperationResponse::fail(
-                request_id,
-                format!("failed to stage AVR8js firmware.hex: {}", e),
-            )),
-        );
-    }
-
-    let staged_elf = if let Some(ref elf) = elf_path {
-        let dest = session_dir.join("firmware.elf");
-        match std::fs::copy(elf, &dest) {
-            Ok(_) => Some(dest),
-            Err(_) => None,
-        }
-    } else {
-        None
-    };
-
-    let manifest = Avr8jsSessionManifest {
-        session_id: session_id.clone(),
-        project_dir: project_dir.to_string_lossy().to_string(),
-        env_name: env_name.clone(),
-        board_id,
-        platform: format!("{:?}", platform),
-        mcu: board.mcu.clone(),
-        f_cpu_hz: board
-            .f_cpu
-            .trim_end_matches('L')
-            .parse::<u32>()
-            .unwrap_or(16_000_000),
-        firmware_hex: staged_hex.to_string_lossy().to_string(),
-        firmware_elf: staged_elf.map(|p| p.to_string_lossy().to_string()),
-        created_at_unix: now_unix(),
-    };
-    let manifest_path = session_dir.join("session.json");
-    if let Err(e) = std::fs::write(
-        &manifest_path,
-        serde_json::to_vec_pretty(&manifest).unwrap_or_default(),
-    ) {
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(OperationResponse::fail(
-                request_id,
-                format!("failed to write AVR8js session manifest: {}", e),
-            )),
-        );
-    }
-    ctx.avr8js_sessions
-        .insert(session_id.clone(), manifest_path);
-
-    if monitor_after {
-        // Headless path: run avr8js in Node.js subprocess, capture UART on stdout
-        let node_path = match find_node() {
-            Ok(p) => p,
-            Err(e) => {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(OperationResponse::fail(request_id, e.to_string())),
-                );
-            }
-        };
-        let avr8js_cache = match ensure_avr8js_npm() {
-            Ok(p) => p,
-            Err(e) => {
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(OperationResponse::fail(request_id, e.to_string())),
-                );
-            }
-        };
-
-        let script_path = session_dir.join("headless.mjs");
-        if let Err(e) = std::fs::write(&script_path, AVR8JS_HEADLESS_MJS) {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("failed to write headless.mjs: {}", e),
-                )),
-            );
-        }
-
-        let avr8js_result = match run_avr8js_headless(
-            &node_path,
-            &script_path,
-            &staged_hex,
-            manifest.f_cpu_hz,
-            &avr8js_cache,
-            RunAvr8jsHeadlessOptions {
-                timeout_secs: monitor_timeout,
-                halt_on_error: halt_on_error.as_deref(),
-                halt_on_success: halt_on_success.as_deref(),
-                expect: expect.as_deref(),
-                show_timestamp,
-                verbose,
-            },
-        )
-        .await
-        {
-            Ok(r) => r,
-            Err(e) => {
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(OperationResponse::fail(request_id, e.to_string())),
-                );
-            }
-        };
-
-        match avr8js_result.outcome {
-            MonitorOutcome::Success(message) => (
-                StatusCode::OK,
-                Json(OperationResponse {
-                    success: true,
-                    request_id,
-                    message: format!("avr8js run succeeded: {}", message),
-                    exit_code: 0,
-                    output_file: Some(output_file),
-                    output_dir,
-                    launch_url: None,
-                    stdout: Some(avr8js_result.stdout),
-                    stderr: Some(avr8js_result.stderr),
-                }),
-            ),
-            MonitorOutcome::Error(message) => (
-                StatusCode::OK,
-                Json(OperationResponse {
-                    success: false,
-                    request_id,
-                    message: format!("avr8js run failed: {}", message),
-                    exit_code: 1,
-                    output_file: Some(output_file),
-                    output_dir,
-                    launch_url: None,
-                    stdout: Some(avr8js_result.stdout),
-                    stderr: Some(avr8js_result.stderr),
-                }),
-            ),
-            MonitorOutcome::Timeout { expect_found } => {
-                let success = expect.is_none() || expect_found;
-                let exit_code = if success { 0 } else { 1 };
-                (
-                    StatusCode::OK,
-                    Json(OperationResponse {
-                        success,
-                        request_id,
-                        message: if success {
-                            "avr8js run completed (timeout)".to_string()
-                        } else {
-                            "avr8js run timed out (expected pattern not found)".to_string()
-                        },
-                        exit_code,
-                        output_file: Some(output_file),
-                        output_dir,
-                        launch_url: None,
-                        stdout: Some(avr8js_result.stdout),
-                        stderr: Some(avr8js_result.stderr),
-                    }),
-                )
-            }
-        }
-    } else {
-        // Browser path: return URL for the avr8js web UI
-        let launch_url = Some(format!(
-            "http://127.0.0.1:{}/emulator/avr8js/{}",
-            ctx.port, session_id
-        ));
-        (
-            StatusCode::OK,
-            Json(OperationResponse {
-                success: true,
-                request_id,
-                message: "deploy complete".to_string(),
-                exit_code: 0,
-                output_file: Some(output_file),
-                output_dir,
-                launch_url,
-                stdout: None,
-                stderr: None,
-            }),
-        )
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Headless avr8js helpers
-// ---------------------------------------------------------------------------
-
-fn find_node() -> fbuild_core::Result<PathBuf> {
-    let node = if cfg!(windows) { "node.exe" } else { "node" };
-    // Route through fbuild-core's `run_command` so the probe spawn is
-    // captured by the daemon's containment group (issue #32). The probe
-    // is short-lived (`node --version`) but a missing binary should
-    // still bubble up the same way.
-    match fbuild_core::subprocess::run_command(&[node, "--version"], None, None, None) {
-        Ok(output) if output.success() => Ok(PathBuf::from(node)),
-        _ => Err(fbuild_core::FbuildError::DeployFailed(
-            "Node.js is required for headless avr8js emulation but 'node' was not found on PATH. \
-             Install Node.js 18+ from https://nodejs.org/"
-                .to_string(),
-        )),
-    }
-}
-
-/// Env-var that forces `ensure_avr8js_npm` to wipe and reinstall the cache
-/// regardless of the integrity marker. Set `FBUILD_REFRESH_EMU_CACHE=1` before
-/// invoking `fbuild test-emu` (or restart the daemon with it in the env) to
-/// recover from a corrupt or partial avr8js install.
-const REFRESH_EMU_CACHE_ENV: &str = "FBUILD_REFRESH_EMU_CACHE";
-
-fn refresh_emu_cache_requested() -> bool {
-    matches!(
-        std::env::var(REFRESH_EMU_CACHE_ENV)
-            .ok()
-            .as_deref()
-            .map(str::trim),
-        Some("1") | Some("true") | Some("TRUE") | Some("yes") | Some("YES")
-    )
-}
-
-fn avr8js_cache_is_intact(cache_dir: &Path) -> bool {
-    // Cheapest proof of a non-corrupt install: both the `avr8js` module dir
-    // *and* its `package.json` must exist. A stale/partial `npm install`
-    // sometimes leaves the directory without the manifest and causes
-    // Node to fail with `ERR_MODULE_NOT_FOUND` at runtime.
-    let module_dir = cache_dir.join("node_modules").join("avr8js");
-    let marker = module_dir.join("package.json");
-    module_dir.is_dir() && marker.is_file()
-}
-
-/// Describes what `prepare_avr8js_cache_for_install` did to the cache dir
-/// before a reinstall attempt. Exposed for unit testing; production code
-/// only needs the side-effect on disk.
-#[derive(Debug, PartialEq, Eq)]
-enum Avr8jsCachePrep {
-    /// Cache was already intact — no-op, reinstall not required.
-    AlreadyIntact,
-    /// No preexisting cache tree; reinstall will populate a fresh dir.
-    NothingToClean,
-    /// `node_modules/` was present but corrupt (missing marker); wiped.
-    WipedNodeModules,
-    /// Force-refresh was requested; entire `cache_dir` was wiped.
-    ForceWiped,
-}
-
-/// Inspect `cache_dir`, wipe corrupt/partial installs, and return what was
-/// done. Does NOT run `npm install`.
-fn prepare_avr8js_cache_for_install(cache_dir: &Path, force_refresh: bool) -> Avr8jsCachePrep {
-    if !force_refresh && avr8js_cache_is_intact(cache_dir) {
-        return Avr8jsCachePrep::AlreadyIntact;
-    }
-
-    if force_refresh && cache_dir.exists() {
-        tracing::info!(
-            "{}=1 set; wiping avr8js cache at {}",
-            REFRESH_EMU_CACHE_ENV,
-            cache_dir.display()
-        );
-        if let Err(e) = std::fs::remove_dir_all(cache_dir) {
-            tracing::warn!(
-                "failed to wipe avr8js cache at {}: {} (continuing with reinstall)",
-                cache_dir.display(),
-                e
-            );
-        }
-        return Avr8jsCachePrep::ForceWiped;
-    }
-
-    if cache_dir.exists() {
-        let node_modules = cache_dir.join("node_modules");
-        if node_modules.exists() {
-            tracing::warn!(
-                "avr8js cache at {} is corrupt (missing node_modules/avr8js/package.json); reinstalling",
-                cache_dir.display()
-            );
-            if let Err(e) = std::fs::remove_dir_all(&node_modules) {
-                tracing::warn!(
-                    "failed to wipe avr8js node_modules at {}: {} (continuing with reinstall)",
-                    node_modules.display(),
-                    e
-                );
-            }
-            return Avr8jsCachePrep::WipedNodeModules;
-        }
-    }
-
-    Avr8jsCachePrep::NothingToClean
-}
-
-fn ensure_avr8js_npm() -> fbuild_core::Result<PathBuf> {
-    let cache_dir = fbuild_paths::get_cache_root().join("avr8js-node");
-    ensure_avr8js_npm_in(&cache_dir, refresh_emu_cache_requested())?;
-    Ok(cache_dir)
-}
-
-/// Populate `cache_dir` with a fresh `node_modules/avr8js` install, wiping
-/// a corrupt or partial install as needed. Split out from `ensure_avr8js_npm`
-/// so unit tests can inject a temporary cache dir without touching env vars.
-fn ensure_avr8js_npm_in(cache_dir: &Path, force_refresh: bool) -> fbuild_core::Result<()> {
-    if prepare_avr8js_cache_for_install(cache_dir, force_refresh) == Avr8jsCachePrep::AlreadyIntact
-    {
-        return Ok(());
-    }
-
-    std::fs::create_dir_all(cache_dir).map_err(|e| {
-        fbuild_core::FbuildError::DeployFailed(format!(
-            "failed to create avr8js cache dir at {}: {}",
-            cache_dir.display(),
-            e
-        ))
-    })?;
-
-    let npm = if cfg!(windows) { "npm.cmd" } else { "npm" };
-    // Route through `run_command` (which spawns via the daemon's
-    // containment group) so an `npm install` killed mid-flight doesn't
-    // leak node processes after the daemon dies. See FastLED/fbuild#32.
-    let cache_dir_str = cache_dir.to_string_lossy().to_string();
-    let output = fbuild_core::subprocess::run_command(
-        &[
-            npm,
-            "install",
-            "--save",
-            "avr8js@0.21.0",
-            "--prefix",
-            &cache_dir_str,
-        ],
-        None,
-        None,
-        None,
-    )
-    .map_err(|e| {
-        fbuild_core::FbuildError::DeployFailed(format!(
-            "failed to launch 'npm' (for `npm install avr8js@0.21.0 --prefix {}`): {}. \
-             Ensure `npm` is installed alongside Node.js and on PATH \
-             (https://nodejs.org/). If npm is installed, set \
-             {}=1 to force a clean reinstall.",
-            cache_dir.display(),
-            e,
-            REFRESH_EMU_CACHE_ENV
-        ))
-    })?;
-    if !output.success() {
-        return Err(fbuild_core::FbuildError::DeployFailed(format!(
-            "`npm install avr8js@0.21.0 --prefix {}` exited with status {}.\n\
-             --- stdout ---\n{}\n--- stderr ---\n{}",
-            cache_dir.display(),
-            output.exit_code,
-            output.stdout.trim_end(),
-            output.stderr.trim_end()
-        )));
-    }
-
-    // Post-install integrity check: guard against npm exiting 0 without
-    // actually extracting the package (rare, but has been seen on Windows
-    // under antivirus interference).
-    if !avr8js_cache_is_intact(cache_dir) {
-        return Err(fbuild_core::FbuildError::DeployFailed(format!(
-            "avr8js npm install at {} reported success but the cache is still \
-             incomplete (missing node_modules/avr8js/package.json). \
-             Try rerunning with {}=1.",
-            cache_dir.display(),
-            REFRESH_EMU_CACHE_ENV
-        )));
-    }
-
-    Ok(())
-}
-
-struct Avr8jsRunResult {
-    outcome: MonitorOutcome,
-    stdout: String,
-    stderr: String,
-    exit_code: Option<i32>,
-}
-
-struct RunAvr8jsHeadlessOptions<'a> {
-    timeout_secs: Option<f64>,
-    halt_on_error: Option<&'a str>,
-    halt_on_success: Option<&'a str>,
-    expect: Option<&'a str>,
-    show_timestamp: bool,
-    verbose: bool,
-}
-
-async fn run_avr8js_headless(
-    node_path: &Path,
-    script_path: &Path,
-    hex_path: &Path,
-    f_cpu_hz: u32,
-    avr8js_cache_dir: &Path,
-    options: RunAvr8jsHeadlessOptions<'_>,
-) -> fbuild_core::Result<Avr8jsRunResult> {
-    // Fail-fast: re-verify the cache is intact before we spawn Node. This
-    // guards against race conditions (cache wiped between `ensure_avr8js_npm`
-    // and this call) and makes the error actionable instead of a cryptic
-    // Node `ERR_MODULE_NOT_FOUND` stack trace (issue #86).
-    if !avr8js_cache_is_intact(avr8js_cache_dir) {
-        tracing::error!(
-            "avr8js cache not populated at {}; aborting (expected \
-             node_modules/avr8js/package.json). Set {}=1 to force reinstall.",
-            avr8js_cache_dir.display(),
-            REFRESH_EMU_CACHE_ENV
-        );
-        return Err(fbuild_core::FbuildError::DeployFailed(format!(
-            "avr8js cache not populated at {} (missing \
-             node_modules/avr8js/package.json). \
-             Set {}=1 to force reinstall.",
-            avr8js_cache_dir.display(),
-            REFRESH_EMU_CACHE_ENV
-        )));
-    }
-
-    // allow-direct-spawn: tokio streaming emulator; blocking NativeProcess unsuitable.
-    let mut cmd = tokio::process::Command::new(node_path);
-    cmd.arg(script_path)
-        .arg("--hex")
-        .arg(hex_path)
-        .arg("--f-cpu")
-        .arg(f_cpu_hz.to_string())
-        .env("NODE_PATH", avr8js_cache_dir.join("node_modules"))
-        .stdin(Stdio::null())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped());
-
-    #[cfg(windows)]
-    {
-        const CREATE_NO_WINDOW: u32 = 0x08000000;
-        cmd.creation_flags(CREATE_NO_WINDOW);
-    }
-
-    if options.verbose {
-        tracing::info!(
-            "avr8js headless: {} {} --hex {} --f-cpu {}",
-            node_path.display(),
-            script_path.display(),
-            hex_path.display(),
-            f_cpu_hz
-        );
-    }
-
-    // Route through containment (#32) so a daemon crash mid-emulation
-    // takes node.exe and any helper processes it spawned with it.
-    let mut child =
-        fbuild_core::containment::tokio_spawn::spawn_contained(&mut cmd).map_err(|e| {
-            fbuild_core::FbuildError::DeployFailed(format!(
-                "failed to launch Node.js for avr8js: {}",
-                e
-            ))
-        })?;
-
-    let stdout = child.stdout.take().ok_or_else(|| {
-        fbuild_core::FbuildError::DeployFailed("failed to capture avr8js stdout".to_string())
-    })?;
-    let stderr = child.stderr.take().ok_or_else(|| {
-        fbuild_core::FbuildError::DeployFailed("failed to capture avr8js stderr".to_string())
-    })?;
-
-    let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<ProcessEvent>();
-    let stdout_task = tokio::spawn(spawn_line_reader(stdout, false, tx.clone()));
-    let stderr_task = tokio::spawn(spawn_line_reader(stderr, true, tx));
-
-    let mut monitor = MonitorState::new(
-        options.timeout_secs,
-        options.halt_on_error,
-        options.halt_on_success,
-        options.expect,
-        options.show_timestamp,
-    );
-    let mut stdout_buf = String::new();
-    let mut stderr_buf = String::new();
-    let mut streams_open = 2usize;
-    let mut child_exit: Option<std::process::ExitStatus> = None;
-    let mut final_outcome: Option<MonitorOutcome> = None;
-
-    loop {
-        if monitor.timed_out() {
-            final_outcome = Some(monitor.timeout_outcome());
-            let _ = child.kill().await;
-            break;
-        }
-
-        let recv_timeout = monitor
-            .remaining()
-            .unwrap_or(std::time::Duration::from_secs(1));
-
-        tokio::select! {
-            status = child.wait(), if child_exit.is_none() => {
-                child_exit = Some(status.map_err(|e| {
-                    fbuild_core::FbuildError::DeployFailed(format!("avr8js wait failed: {}", e))
-                })?);
-                if streams_open == 0 {
-                    break;
-                }
-            }
-            maybe_event = tokio::time::timeout(recv_timeout, rx.recv()) => {
-                match maybe_event {
-                    Ok(Some(ProcessEvent::Line(line))) => {
-                        let target = if line.is_stderr { &mut stderr_buf } else { &mut stdout_buf };
-                        target.push_str(&line.line);
-                        target.push('\n');
-
-                        if let Some(outcome) = monitor.process_line(&line.line) {
-                            final_outcome = Some(outcome);
-                            let _ = child.kill().await;
-                            break;
-                        }
-                    }
-                    Ok(Some(ProcessEvent::StreamClosed)) => {
-                        streams_open = streams_open.saturating_sub(1);
-                        if streams_open == 0 && child_exit.is_some() {
-                            break;
-                        }
-                    }
-                    Ok(None) => {
-                        if child_exit.is_some() {
-                            break;
-                        }
-                    }
-                    Err(_) => {
-                        final_outcome = Some(monitor.timeout_outcome());
-                        let _ = child.kill().await;
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    if child_exit.is_none() {
-        child_exit = Some(child.wait().await.map_err(|e| {
-            fbuild_core::FbuildError::DeployFailed(format!("avr8js wait failed: {}", e))
-        })?);
-    }
-
-    let _ = stdout_task.await;
-    let _ = stderr_task.await;
-
-    let outcome = if let Some(outcome) = final_outcome {
-        outcome
-    } else if let Some(status) = child_exit {
-        if status.success() {
-            if options.expect.is_some() && !monitor.expect_found() {
-                MonitorOutcome::Error(
-                    "avr8js exited before the expected pattern was found".to_string(),
-                )
-            } else {
-                MonitorOutcome::Success("avr8js exited normally".to_string())
-            }
-        } else {
-            MonitorOutcome::Error(format!(
-                "avr8js exited with code {}",
-                status.code().unwrap_or(-1)
-            ))
-        }
-    } else {
-        MonitorOutcome::Error("avr8js exited unexpectedly".to_string())
-    };
-
-    Ok(Avr8jsRunResult {
-        outcome,
-        stdout: stdout_buf,
-        stderr: stderr_buf,
-        exit_code: child_exit.and_then(|s| s.code()),
-    })
-}
-
-fn qemu_session_dir(project_dir: &Path, env_name: &str) -> PathBuf {
-    fbuild_paths::get_project_fbuild_dir(project_dir)
-        .join("emulators")
-        .join("qemu")
-        .join(env_name)
-        .join(uuid::Uuid::new_v4().to_string())
-}
-
-fn build_linux_macos_qemu_hint(err: &str) -> String {
-    if cfg!(any(target_os = "linux", target_os = "macos")) {
-        format!(
-            "{}. On Linux/macOS, ensure QEMU runtime deps are installed: libgcrypt, glib2, pixman, SDL2, and libslirp.",
-            err
-        )
-    } else {
-        err.to_string()
-    }
-}
-
-fn resolve_esp32_toolchain_gcc_path(
-    project_dir: &Path,
-    mcu_config: &fbuild_build::esp32::mcu_config::Esp32McuConfig,
-) -> fbuild_core::Result<PathBuf> {
-    let platform = fbuild_packages::library::Esp32Platform::new(project_dir);
-    Package::ensure_installed(&platform)?;
-
-    let is_riscv = mcu_config.is_riscv();
-    let prefix = mcu_config.toolchain_prefix();
-    let toolchain_name = if is_riscv {
-        "toolchain-riscv32-esp"
-    } else {
-        "toolchain-xtensa-esp-elf"
-    };
-
-    let toolchain = match platform.get_toolchain_metadata_url(is_riscv) {
-        Ok(metadata_url) => {
-            let cache = fbuild_packages::Cache::new(project_dir);
-            let cache_dir = cache.toolchains_dir().join(toolchain_name);
-            match fbuild_packages::toolchain::esp32_metadata::resolve_toolchain_url_sync(
-                &metadata_url,
-                toolchain_name,
-                &cache_dir,
-            ) {
-                Ok(resolved) => fbuild_packages::toolchain::Esp32Toolchain::from_resolved(
-                    project_dir,
-                    &resolved.url,
-                    resolved.sha256.as_deref(),
-                    is_riscv,
-                    &prefix,
-                ),
-                Err(_) => {
-                    fbuild_packages::toolchain::Esp32Toolchain::new(project_dir, is_riscv, &prefix)
-                }
-            }
-        }
-        Err(_) => fbuild_packages::toolchain::Esp32Toolchain::new(project_dir, is_riscv, &prefix),
-    };
-
-    let _ = Package::ensure_installed(&toolchain)?;
-    Ok(toolchain.get_gcc_path())
-}
-
-#[cfg(windows)]
-fn apply_windows_process_flags(cmd: &mut tokio::process::Command, exe_path: &Path) {
-    const CREATE_NO_WINDOW: u32 = 0x08000000;
-    cmd.creation_flags(CREATE_NO_WINDOW);
-
-    let current_path = std::env::var("PATH").unwrap_or_default();
-    if let Ok(path_env) =
-        fbuild_packages::toolchain::build_windows_qemu_path_env(exe_path, &current_path)
-    {
-        cmd.env("PATH", path_env);
-    }
-}
-
-#[cfg(not(windows))]
-fn apply_windows_process_flags(_cmd: &mut tokio::process::Command, _exe_path: &Path) {}
-
-async fn spawn_line_reader(
-    stream: impl tokio::io::AsyncRead + Unpin + Send + 'static,
-    is_stderr: bool,
-    tx: tokio::sync::mpsc::UnboundedSender<ProcessEvent>,
-) {
-    let mut lines = BufReader::new(stream).lines();
-    while let Ok(Some(line)) = lines.next_line().await {
-        let _ = tx.send(ProcessEvent::Line(ProcessLine { is_stderr, line }));
-    }
-    let _ = tx.send(ProcessEvent::StreamClosed);
-}
-
-async fn run_qemu_process(
-    qemu_path: &Path,
-    args: &[String],
-    options: RunQemuOptions<'_>,
-) -> fbuild_core::Result<QemuRunResult> {
-    // allow-direct-spawn: tokio streaming QEMU emulator; blocking NativeProcess unsuitable.
-    let mut cmd = tokio::process::Command::new(qemu_path);
-    cmd.args(args)
-        .stdin(Stdio::null())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped());
-    apply_windows_process_flags(&mut cmd, qemu_path);
-
-    let label = options.process_label;
-    if options.verbose {
-        tracing::info!("{}: {} {}", label, qemu_path.display(), args.join(" "));
-    }
-
-    // Route through containment (#32). QEMU is a long-running process
-    // — a daemon hard-kill mid-emulation must not leave `qemu-system-*`
-    // or its `conhost.exe` wrapper behind.
-    let mut child =
-        fbuild_core::containment::tokio_spawn::spawn_contained(&mut cmd).map_err(|e| {
-            fbuild_core::FbuildError::DeployFailed(build_linux_macos_qemu_hint(&format!(
-                "failed to launch {} at {}: {}",
-                label,
-                qemu_path.display(),
-                e
-            )))
-        })?;
-
-    let stdout = child.stdout.take().ok_or_else(|| {
-        fbuild_core::FbuildError::DeployFailed(format!("failed to capture {} stdout", label))
-    })?;
-    let stderr = child.stderr.take().ok_or_else(|| {
-        fbuild_core::FbuildError::DeployFailed(format!("failed to capture {} stderr", label))
-    })?;
-
-    let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<ProcessEvent>();
-    let stdout_task = tokio::spawn(spawn_line_reader(stdout, false, tx.clone()));
-    let stderr_task = tokio::spawn(spawn_line_reader(stderr, true, tx));
-
-    let mut monitor = MonitorState::new(
-        options.timeout_secs,
-        options.halt_on_error,
-        options.halt_on_success,
-        options.expect,
-        options.show_timestamp,
-    );
-    let mut crash_decoder =
-        fbuild_serial::crash_decoder::CrashDecoder::new(options.elf_path, options.addr2line_path);
-    let mut stdout_buf = String::new();
-    let mut stderr_buf = String::new();
-    let mut synthetic_buf = String::new();
-    let mut streams_open = 2usize;
-    let mut child_exit: Option<std::process::ExitStatus> = None;
-    let mut final_outcome: Option<MonitorOutcome> = None;
-
-    loop {
-        if monitor.timed_out() {
-            final_outcome = Some(monitor.timeout_outcome());
-            let _ = child.kill().await;
-            break;
-        }
-
-        let recv_timeout = monitor
-            .remaining()
-            .unwrap_or(std::time::Duration::from_secs(1));
-
-        tokio::select! {
-            status = child.wait(), if child_exit.is_none() => {
-                child_exit = Some(status.map_err(|e| {
-                    fbuild_core::FbuildError::DeployFailed(format!("{} wait failed: {}", label, e))
-                })?);
-                if streams_open == 0 {
-                    break;
-                }
-            }
-            maybe_event = tokio::time::timeout(recv_timeout, rx.recv()) => {
-                match maybe_event {
-                    Ok(Some(ProcessEvent::Line(line))) => {
-                        let target = if line.is_stderr { &mut stderr_buf } else { &mut stdout_buf };
-                        target.push_str(&line.line);
-                        target.push('\n');
-
-                        if let Some(outcome) = monitor.process_line(&line.line) {
-                            final_outcome = Some(outcome);
-                            let _ = child.kill().await;
-                            break;
-                        }
-
-                        if let Some(decoded_lines) = crash_decoder.process_line(&line.line) {
-                            for decoded in decoded_lines {
-                                synthetic_buf.push_str(&decoded);
-                                synthetic_buf.push('\n');
-                                if let Some(outcome) = monitor.process_line(&decoded) {
-                                    final_outcome = Some(outcome);
-                                    let _ = child.kill().await;
-                                    break;
-                                }
-                            }
-                            if final_outcome.is_some() {
-                                break;
-                            }
-                        }
-                    }
-                    Ok(Some(ProcessEvent::StreamClosed)) => {
-                        streams_open = streams_open.saturating_sub(1);
-                        if streams_open == 0 && child_exit.is_some() {
-                            break;
-                        }
-                    }
-                    Ok(None) => {
-                        if child_exit.is_some() {
-                            break;
-                        }
-                    }
-                    Err(_) => {
-                        final_outcome = Some(monitor.timeout_outcome());
-                        let _ = child.kill().await;
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    if child_exit.is_none() {
-        child_exit = Some(child.wait().await.map_err(|e| {
-            fbuild_core::FbuildError::DeployFailed(format!("{} wait failed: {}", label, e))
-        })?);
-    }
-
-    let _ = stdout_task.await;
-    let _ = stderr_task.await;
-
-    if !synthetic_buf.is_empty() {
-        stdout_buf.push_str(&synthetic_buf);
-    }
-
-    let outcome = if let Some(outcome) = final_outcome {
-        outcome
-    } else if let Some(status) = child_exit {
-        if status.success() {
-            if options.expect.is_some() && !monitor.expect_found() {
-                MonitorOutcome::Error(format!(
-                    "{} exited before the expected pattern was found",
-                    label
-                ))
-            } else {
-                MonitorOutcome::Success(format!("{} exited normally", label))
-            }
-        } else {
-            MonitorOutcome::Error(format!(
-                "{} exited with code {}",
-                label,
-                status.code().unwrap_or(-1)
-            ))
-        }
-    } else {
-        MonitorOutcome::Error(format!("{} exited unexpectedly", label))
-    };
-
-    Ok(QemuRunResult {
-        outcome,
-        stdout: stdout_buf,
-        stderr: stderr_buf,
-        exit_code: child_exit.and_then(|s| s.code()),
-    })
-}
-
-pub async fn deploy_qemu(
-    _ctx: Arc<DaemonContext>,
-    req: DeployQemuRequest,
-) -> (StatusCode, Json<OperationResponse>) {
-    let DeployQemuRequest {
-        request_id,
-        project_dir,
-        env_name,
-        board_id,
-        platform,
-        firmware_path,
-        elf_path,
-        output_file,
-        output_dir,
-        monitor_timeout,
-        qemu_timeout_secs,
-        halt_on_error,
-        halt_on_success,
-        expect,
-        show_timestamp,
-        verbose,
-        board_overrides,
-    } = req;
-
-    if platform != fbuild_core::Platform::Espressif32 {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                "QEMU deploy target is currently supported only for ESP32-family boards"
-                    .to_string(),
-            )),
-        );
-    }
-    if firmware_path.extension().and_then(|ext| ext.to_str()) != Some("bin") {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                format!(
-                    "QEMU deploy target requires firmware.bin, got '{}'",
-                    firmware_path.display()
-                ),
-            )),
-        );
-    }
-
-    let board = match fbuild_config::BoardConfig::from_board_id(&board_id, &board_overrides) {
-        Ok(board) => board,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("failed to load board config for QEMU: {}", e),
-                )),
-            );
-        }
-    };
-    if !is_qemu_supported_esp32_mcu(&board.mcu) {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                format!(
-                    "native QEMU deploy currently supports ESP32, ESP32-S3 (Xtensa) and \
-                     ESP32-C3, ESP32-C6, ESP32-H2 (RISC-V) boards, got '{}'",
-                    board.mcu
-                ),
-            )),
-        );
-    }
-
-    let mcu_config = match fbuild_build::esp32::mcu_config::get_mcu_config(&board.mcu) {
-        Ok(cfg) => cfg,
-        Err(e) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("failed to load MCU config for QEMU: {}", e),
-                )),
-            );
-        }
-    };
-
-    let effective_flash_mode = board
-        .flash_mode
-        .as_deref()
-        .unwrap_or(mcu_config.default_flash_mode());
-    if !effective_flash_mode.eq_ignore_ascii_case("dio") {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                format!(
-                    "QEMU requires a DIO-compatible flash image; effective flash mode is '{}'",
-                    effective_flash_mode
-                ),
-            )),
-        );
-    }
-
-    let flash_size_bytes = match fbuild_deploy::esp32::resolve_qemu_flash_size_bytes(
-        &board,
-        mcu_config.default_flash_size(),
-    ) {
-        Ok(size) => size,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(OperationResponse::fail(request_id, e.to_string())),
-            );
-        }
-    };
-
-    let qemu = match resolve_esp_qemu_for_mcu(&project_dir, &board.mcu) {
-        Ok(path) => path,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(OperationResponse::fail(
-                    request_id,
-                    build_linux_macos_qemu_hint(&e.to_string()),
-                )),
-            );
-        }
-    };
-
-    let session_dir = qemu_session_dir(&project_dir, &env_name);
-    if let Err(e) = std::fs::create_dir_all(&session_dir) {
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(OperationResponse::fail(
-                request_id,
-                format!("failed to create QEMU session dir: {}", e),
-            )),
-        );
-    }
-    let flash_image = session_dir.join("qemu_flash.bin");
-    // Only apply the ESP32-S3 ADC calibration patch for S3 variants.
-    let elf_for_adc_patch = if board.mcu.eq_ignore_ascii_case("esp32s3") {
-        elf_path.as_deref()
-    } else {
-        None
-    };
-    if let Err(e) = fbuild_deploy::esp32::create_qemu_flash_image(
-        &firmware_path,
-        &flash_image,
-        flash_size_bytes,
-        mcu_config.bootloader_offset(),
-        mcu_config.partitions_offset(),
-        mcu_config.firmware_offset(),
-        elf_for_adc_patch,
-    ) {
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(OperationResponse::fail(
-                request_id,
-                format!("failed to create QEMU flash image: {}", e),
-            )),
-        );
-    }
-
-    let args = fbuild_deploy::esp32::build_qemu_args(
-        &board.mcu,
-        &flash_image,
-        board.qemu_esp32_psram_config(),
-    );
-    let addr2line_path = elf_path.as_ref().and_then(|_| {
-        resolve_esp32_toolchain_gcc_path(&project_dir, &mcu_config)
-            .ok()
-            .and_then(|gcc| fbuild_serial::crash_decoder::derive_addr2line_path(&gcc))
-    });
-
-    let timeout_secs = monitor_timeout.or(Some(qemu_timeout_secs as f64));
-    let qemu_result = match run_qemu_process(
-        &qemu,
-        &args,
-        RunQemuOptions {
-            elf_path,
-            addr2line_path,
-            timeout_secs,
-            halt_on_error: halt_on_error.as_deref(),
-            halt_on_success: halt_on_success.as_deref(),
-            expect: expect.as_deref(),
-            show_timestamp,
-            verbose,
-            process_label: "QEMU",
-        },
-    )
-    .await
-    {
-        Ok(result) => result,
-        Err(e) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(request_id, e.to_string())),
-            );
-        }
-    };
-
-    let real_exit_code = qemu_result.exit_code;
-    match qemu_result.outcome {
-        MonitorOutcome::Success(message) => (
-            StatusCode::OK,
-            Json(OperationResponse {
-                success: true,
-                request_id,
-                message: format!("QEMU run succeeded: {}", message),
-                exit_code: real_exit_code.unwrap_or(0),
-                output_file: Some(output_file),
-                output_dir,
-                launch_url: None,
-                stdout: Some(qemu_result.stdout),
-                stderr: Some(qemu_result.stderr),
-            }),
-        ),
-        MonitorOutcome::Error(message) => (
-            StatusCode::OK,
-            Json(OperationResponse {
-                success: false,
-                request_id,
-                message: format!("QEMU run failed: {}", message),
-                exit_code: real_exit_code.unwrap_or(1),
-                output_file: Some(output_file),
-                output_dir,
-                launch_url: None,
-                stdout: Some(qemu_result.stdout),
-                stderr: Some(qemu_result.stderr),
-            }),
-        ),
-        MonitorOutcome::Timeout { expect_found } => {
-            let success = expect.is_none() || expect_found;
-            let exit_code = real_exit_code.unwrap_or(if success { 0 } else { 1 });
-            (
-                StatusCode::OK,
-                Json(OperationResponse {
-                    success,
-                    request_id,
-                    message: if success {
-                        "QEMU run completed (timeout)".to_string()
-                    } else {
-                        "QEMU run timed out (expected pattern not found)".to_string()
-                    },
-                    exit_code,
-                    output_file: Some(output_file),
-                    output_dir,
-                    launch_url: None,
-                    stdout: Some(qemu_result.stdout),
-                    stderr: Some(qemu_result.stderr),
-                }),
-            )
-        }
-    }
-}
-
-pub async fn avr8js_page(
-    AxumPath(session_id): AxumPath<String>,
-    State(ctx): State<Arc<DaemonContext>>,
-) -> impl IntoResponse {
-    match load_session_manifest(&ctx, &session_id) {
-        Ok(_) => Html(render_page(&session_id)).into_response(),
-        Err(e) => (StatusCode::NOT_FOUND, e.to_string()).into_response(),
-    }
-}
-
-pub async fn avr8js_app_js() -> impl IntoResponse {
-    (
-        [(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static("application/javascript; charset=utf-8"),
-        )],
-        AVR8JS_APP_JS,
-    )
-}
-
-pub async fn avr8js_session_json(
-    AxumPath(session_id): AxumPath<String>,
-    State(ctx): State<Arc<DaemonContext>>,
-) -> impl IntoResponse {
-    match load_session_manifest(&ctx, &session_id) {
-        Ok(manifest) => (
-            StatusCode::OK,
-            Json(Avr8jsSessionResponse {
-                session_id: manifest.session_id.clone(),
-                env_name: manifest.env_name,
-                board_id: manifest.board_id,
-                platform: manifest.platform,
-                mcu: manifest.mcu,
-                f_cpu_hz: manifest.f_cpu_hz,
-                firmware_hex_url: format!("/api/emulator/avr8js/{}/firmware.hex", session_id),
-            }),
-        )
-            .into_response(),
-        Err(e) => (StatusCode::NOT_FOUND, e.to_string()).into_response(),
-    }
-}
-
-pub async fn avr8js_firmware_hex(
-    AxumPath(session_id): AxumPath<String>,
-    State(ctx): State<Arc<DaemonContext>>,
-) -> impl IntoResponse {
-    match load_session_manifest(&ctx, &session_id) {
-        Ok(manifest) => match std::fs::read_to_string(&manifest.firmware_hex) {
-            Ok(hex) => (
-                [(
-                    header::CONTENT_TYPE,
-                    HeaderValue::from_static("text/plain; charset=utf-8"),
-                )],
-                hex,
-            )
-                .into_response(),
-            Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()).into_response(),
-        },
-        Err(e) => (StatusCode::NOT_FOUND, e.to_string()).into_response(),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// EmulatorRunner abstraction (Issue #23)
-// ---------------------------------------------------------------------------
-
-use fbuild_core::emulator::{EmulatorArtifactBundle, EmulatorOutcome, EmulatorRunResult};
-
-/// Configuration for an emulator test run (user-facing options).
-pub struct EmulatorRunConfig {
-    pub firmware_path: PathBuf,
-    pub elf_path: Option<PathBuf>,
-    /// Structured artifact bundle for runner validation.
-    pub artifact_bundle: EmulatorArtifactBundle,
-    pub timeout: Option<f64>,
-    pub halt_on_error: Option<String>,
-    pub halt_on_success: Option<String>,
-    pub expect: Option<String>,
-    pub show_timestamp: bool,
-    pub verbose: bool,
-}
-
-/// Convert a `MonitorOutcome` into an `EmulatorOutcome`.
-fn monitor_outcome_to_emulator(outcome: MonitorOutcome, exit_code: Option<i32>) -> EmulatorOutcome {
-    match outcome {
-        MonitorOutcome::Success(msg) => EmulatorOutcome::Passed(msg),
-        MonitorOutcome::Error(msg) => {
-            // Heuristic: if the process crashed (non-zero exit with crash signature)
-            if let Some(code) = exit_code {
-                if code != 0 && (msg.contains("abort()") || msg.contains("Guru Meditation")) {
-                    return EmulatorOutcome::Crashed(msg);
-                }
-            }
-            EmulatorOutcome::Failed(msg)
-        }
-        MonitorOutcome::Timeout { expect_found } => EmulatorOutcome::TimedOut { expect_found },
-    }
-}
-
-/// Abstraction over emulator backends. Each implementation knows how to set up
-/// and execute a specific emulator (QEMU, avr8js, etc.).
-#[async_trait::async_trait]
-pub trait EmulatorRunner: Send + Sync {
-    /// Human-readable name of this runner (e.g. "QEMU ESP32-S3", "avr8js ATmega328P").
-    fn name(&self) -> &str;
-
-    /// Run the emulator with the given configuration.
-    async fn run(&self, config: &EmulatorRunConfig) -> fbuild_core::Result<EmulatorRunResult>;
-}
-
-/// QEMU-based emulator runner for ESP32-family boards.
-pub struct QemuRunner {
-    project_dir: PathBuf,
-    env_name: String,
-    board: fbuild_config::BoardConfig,
-    display_name: String,
-}
-
-impl QemuRunner {
-    pub fn new(project_dir: PathBuf, env_name: String, board: fbuild_config::BoardConfig) -> Self {
-        let display_name = format!("QEMU {}", board.mcu.to_uppercase());
-        Self {
-            project_dir,
-            env_name,
-            board,
-            display_name,
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl EmulatorRunner for QemuRunner {
-    fn name(&self) -> &str {
-        &self.display_name
-    }
-
-    async fn run(&self, config: &EmulatorRunConfig) -> fbuild_core::Result<EmulatorRunResult> {
-        if let Err(msg) = config
-            .artifact_bundle
-            .validate_for(fbuild_core::emulator::RunnerKind::QemuEsp32)
-        {
-            return Err(fbuild_core::FbuildError::DeployFailed(msg));
-        }
-        let mcu_config = fbuild_build::esp32::mcu_config::get_mcu_config(&self.board.mcu)?;
-
-        let effective_flash_mode = self
-            .board
-            .flash_mode
-            .as_deref()
-            .unwrap_or(mcu_config.default_flash_mode());
-        if !effective_flash_mode.eq_ignore_ascii_case("dio") {
-            return Ok(EmulatorRunResult {
-                outcome: EmulatorOutcome::Unsupported(format!(
-                    "QEMU requires DIO flash mode; effective mode is '{}'",
-                    effective_flash_mode
-                )),
-                stdout: String::new(),
-                stderr: String::new(),
-                command_line: String::new(),
-                exit_code: None,
-            });
-        }
-
-        let flash_size_bytes = fbuild_deploy::esp32::resolve_qemu_flash_size_bytes(
-            &self.board,
-            mcu_config.default_flash_size(),
-        )?;
-
-        let qemu = resolve_esp_qemu_for_mcu(&self.project_dir, &self.board.mcu)?;
-
-        let session_dir = qemu_session_dir(&self.project_dir, &self.env_name);
-        std::fs::create_dir_all(&session_dir)?;
-
-        let flash_image = session_dir.join("qemu_flash.bin");
-
-        // Only apply the ESP32-S3 ADC calibration patch for S3 variants.
-        let elf_for_adc_patch = if self.board.mcu.eq_ignore_ascii_case("esp32s3") {
-            config.elf_path.as_deref()
-        } else {
-            None
-        };
-        fbuild_deploy::esp32::create_qemu_flash_image(
-            &config.firmware_path,
-            &flash_image,
-            flash_size_bytes,
-            mcu_config.bootloader_offset(),
-            mcu_config.partitions_offset(),
-            mcu_config.firmware_offset(),
-            elf_for_adc_patch,
-        )?;
-
-        let args = fbuild_deploy::esp32::build_qemu_args(
-            &self.board.mcu,
-            &flash_image,
-            self.board.qemu_esp32_psram_config(),
-        );
-        let addr2line_path = config.elf_path.as_ref().and_then(|_| {
-            resolve_esp32_toolchain_gcc_path(&self.project_dir, &mcu_config)
-                .ok()
-                .and_then(|gcc| fbuild_serial::crash_decoder::derive_addr2line_path(&gcc))
-        });
-
-        let command_line = format!("{} {}", qemu.display(), args.join(" "));
-
-        let qemu_result = run_qemu_process(
-            &qemu,
-            &args,
-            RunQemuOptions {
-                elf_path: config.elf_path.clone(),
-                addr2line_path,
-                timeout_secs: config.timeout,
-                halt_on_error: config.halt_on_error.as_deref(),
-                halt_on_success: config.halt_on_success.as_deref(),
-                expect: config.expect.as_deref(),
-                show_timestamp: config.show_timestamp,
-                verbose: config.verbose,
-                process_label: "QEMU",
-            },
-        )
-        .await?;
-
-        let exit_code = qemu_result.exit_code;
-        let outcome = monitor_outcome_to_emulator(qemu_result.outcome, exit_code);
-
-        Ok(EmulatorRunResult {
-            outcome,
-            stdout: qemu_result.stdout,
-            stderr: qemu_result.stderr,
-            command_line,
-            exit_code,
-        })
-    }
-}
-
-/// AVR8js-based emulator runner for ATmega328P (headless Node.js).
-pub struct Avr8jsRunner {
-    board: fbuild_config::BoardConfig,
-}
-
-impl Avr8jsRunner {
-    pub fn new(board: fbuild_config::BoardConfig) -> Self {
-        Self { board }
-    }
-}
-
-#[async_trait::async_trait]
-impl EmulatorRunner for Avr8jsRunner {
-    fn name(&self) -> &str {
-        "avr8js ATmega328P"
-    }
-
-    async fn run(&self, config: &EmulatorRunConfig) -> fbuild_core::Result<EmulatorRunResult> {
-        if let Err(msg) = config
-            .artifact_bundle
-            .validate_for(fbuild_core::emulator::RunnerKind::Avr8js)
-        {
-            return Err(fbuild_core::FbuildError::DeployFailed(msg));
-        }
-        let node_path = find_node()?;
-        let avr8js_cache = ensure_avr8js_npm()?;
-
-        let session_dir = tempfile::TempDir::new()?;
-        let script_path = session_dir.path().join("headless.mjs");
-        std::fs::write(&script_path, AVR8JS_HEADLESS_MJS)?;
-
-        let f_cpu_hz: u32 = self
-            .board
-            .f_cpu
-            .trim_end_matches('L')
-            .parse()
-            .unwrap_or(16_000_000);
-
-        let command_line = format!(
-            "{} {} --hex {} --f-cpu {}",
-            node_path.display(),
-            script_path.display(),
-            config.firmware_path.display(),
-            f_cpu_hz
-        );
-
-        let avr8js_result = run_avr8js_headless(
-            &node_path,
-            &script_path,
-            &config.firmware_path,
-            f_cpu_hz,
-            &avr8js_cache,
-            RunAvr8jsHeadlessOptions {
-                timeout_secs: config.timeout,
-                halt_on_error: config.halt_on_error.as_deref(),
-                halt_on_success: config.halt_on_success.as_deref(),
-                expect: config.expect.as_deref(),
-                show_timestamp: config.show_timestamp,
-                verbose: config.verbose,
-            },
-        )
-        .await?;
-
-        let exit_code = avr8js_result.exit_code;
-        let outcome = monitor_outcome_to_emulator(avr8js_result.outcome, exit_code);
-
-        Ok(EmulatorRunResult {
-            outcome,
-            stdout: avr8js_result.stdout,
-            stderr: avr8js_result.stderr,
-            command_line,
-            exit_code,
-        })
-    }
-}
-
-// ---------------------------------------------------------------------------
-// SimAVR native runner (Issue #24)
-// ---------------------------------------------------------------------------
-
-/// Find the `simavr` binary on PATH.
-///
-/// SimAVR is a native AVR simulator. Install via:
-/// - Linux: `apt install simavr` or build from source
-/// - macOS: `brew install simavr`
-/// - Windows: build from source (MSYS2/MinGW) — limited support
-fn find_simavr() -> fbuild_core::Result<PathBuf> {
-    let simavr = if cfg!(windows) {
-        "simavr.exe"
-    } else {
-        "simavr"
-    };
-    // Try running simavr to verify it exists; route through containment
-    // (issue #32). This is a short-lived probe so the containment
-    // difference is purely consistency.
-    match fbuild_core::subprocess::run_command(&[simavr, "--help"], None, None, None) {
-        Ok(_) => Ok(PathBuf::from(simavr)),
-        Err(_) => {
-            let install_hint = if cfg!(target_os = "linux") {
-                "Install via: apt install simavr (Debian/Ubuntu) or your distro's package manager"
-            } else if cfg!(target_os = "macos") {
-                "Install via: brew install simavr"
-            } else {
-                "SimAVR has limited Windows support. Build from source via MSYS2/MinGW, \
-                 or use --emulator avr8js for ATmega328P boards instead"
-            };
-            Err(fbuild_core::FbuildError::DeployFailed(format!(
-                "simavr is required but '{}' was not found on PATH. {}",
-                simavr, install_hint
-            )))
-        }
-    }
-}
-
-/// SimAVR-based native emulator runner for AVR boards.
-///
-/// Supports any AVR board that advertises `simavr` in its debug_tools.
-/// Primary targets: ATmega328P (Uno), ATmega32U4 (Leonardo), ATmega2560 (Mega).
-/// Consumes `firmware.elf` directly.
-pub struct SimavrRunner {
-    board: fbuild_config::BoardConfig,
-}
-
-impl SimavrRunner {
-    pub fn new(board: fbuild_config::BoardConfig) -> Self {
-        Self { board }
-    }
-}
-
-#[async_trait::async_trait]
-impl EmulatorRunner for SimavrRunner {
-    fn name(&self) -> &str {
-        "simavr"
-    }
-
-    async fn run(&self, config: &EmulatorRunConfig) -> fbuild_core::Result<EmulatorRunResult> {
-        if let Err(msg) = config
-            .artifact_bundle
-            .validate_for(fbuild_core::emulator::RunnerKind::Simavr)
-        {
-            return Err(fbuild_core::FbuildError::DeployFailed(msg));
-        }
-        let simavr_path = find_simavr()?;
-
-        // simavr requires an ELF file
-        let elf_path = config.elf_path.as_ref().ok_or_else(|| {
-            fbuild_core::FbuildError::DeployFailed(
-                "simavr requires firmware.elf but no ELF path was produced by the build"
-                    .to_string(),
-            )
-        })?;
-
-        let f_cpu_hz: u32 = self
-            .board
-            .f_cpu
-            .trim_end_matches('L')
-            .parse()
-            .unwrap_or(16_000_000);
-
-        let mcu = self.board.mcu.to_lowercase();
-
-        let args: Vec<String> = vec![
-            "-m".to_string(),
-            mcu.clone(),
-            "-f".to_string(),
-            f_cpu_hz.to_string(),
-            elf_path.display().to_string(),
-        ];
-
-        let command_line = format!("{} {}", simavr_path.display(), args.join(" "));
-
-        // Reuse the generic process runner (run_qemu_process) with no crash decoder
-        let result = run_qemu_process(
-            &simavr_path,
-            &args,
-            RunQemuOptions {
-                elf_path: None,       // no ESP32 crash decoder needed
-                addr2line_path: None, // no addr2line for AVR in this path
-                timeout_secs: config.timeout,
-                halt_on_error: config.halt_on_error.as_deref(),
-                halt_on_success: config.halt_on_success.as_deref(),
-                expect: config.expect.as_deref(),
-                show_timestamp: config.show_timestamp,
-                verbose: config.verbose,
-                process_label: "simavr",
-            },
-        )
-        .await?;
-
-        let exit_code = result.exit_code;
-        let outcome = monitor_outcome_to_emulator(result.outcome, exit_code);
-
-        Ok(EmulatorRunResult {
-            outcome,
-            stdout: result.stdout,
-            stderr: result.stderr,
-            command_line,
-            exit_code,
-        })
-    }
-}
-
-/// Check whether a given MCU is supported by the QEMU runner.
-///
-/// Supported MCUs:
-/// - Xtensa (`qemu-system-xtensa`): `esp32`, `esp32s3`
-/// - RISC-V (`qemu-system-riscv32`): `esp32c3`, `esp32c6`, `esp32h2`
-fn is_qemu_supported_esp32_mcu(mcu: &str) -> bool {
-    fbuild_packages::toolchain::EspQemuArch::for_mcu(mcu).is_some()
-}
-
-/// Resolve the Espressif QEMU executable appropriate for the given MCU.
-///
-/// Picks `qemu-system-xtensa` for ESP32/ESP32-S3 and `qemu-system-riscv32`
-/// for ESP32-C3/C6/H2. Returns the resolved binary path (downloading into
-/// the managed fbuild cache if required).
-fn resolve_esp_qemu_for_mcu(project_dir: &Path, mcu: &str) -> fbuild_core::Result<PathBuf> {
-    let arch = fbuild_packages::toolchain::EspQemuArch::for_mcu(mcu).ok_or_else(|| {
-        fbuild_core::FbuildError::DeployFailed(format!(
-            "no QEMU backend available for MCU '{}'",
-            mcu
-        ))
-    })?;
-    let pkg = fbuild_packages::toolchain::EspQemu::new(project_dir, arch)?;
-    pkg.resolve_executable()
-}
-
-/// Fail fast if the board's flash mode is incompatible with QEMU (DIO only).
-fn check_qemu_flash_mode(board: &fbuild_config::BoardConfig) -> fbuild_core::Result<()> {
-    let mcu_config = fbuild_build::esp32::mcu_config::get_mcu_config(&board.mcu)?;
-    let effective_flash_mode = board
-        .flash_mode
-        .as_deref()
-        .unwrap_or(mcu_config.default_flash_mode());
-    if !effective_flash_mode.eq_ignore_ascii_case("dio") {
-        return Err(fbuild_core::FbuildError::DeployFailed(format!(
-            "QEMU requires DIO flash mode; board '{}' uses '{}'",
-            board.name, effective_flash_mode
-        )));
-    }
-    Ok(())
-}
-
-/// Select the appropriate emulator runner based on platform, MCU, and optional
-/// explicit emulator choice.
-///
-/// Returns `Err` with `EmulatorOutcome::Unsupported` information if no runner
-/// matches.
-pub fn select_runner(
-    project_dir: &Path,
-    env_name: &str,
-    platform: fbuild_core::Platform,
-    board_id: &str,
-    board_overrides: &HashMap<String, String>,
-    emulator: Option<&str>,
-) -> fbuild_core::Result<Box<dyn EmulatorRunner>> {
-    let board = fbuild_config::BoardConfig::from_board_id(board_id, board_overrides)?;
-
-    if let Some(explicit) = emulator {
-        return match explicit {
-            "qemu" => {
-                if platform != fbuild_core::Platform::Espressif32 {
-                    return Err(fbuild_core::FbuildError::DeployFailed(
-                        "QEMU runner is only supported for ESP32-family boards".to_string(),
-                    ));
-                }
-                if !is_qemu_supported_esp32_mcu(&board.mcu) {
-                    return Err(fbuild_core::FbuildError::DeployFailed(format!(
-                        "QEMU runner currently supports ESP32, ESP32-S3 (Xtensa) and \
-                         ESP32-C3, ESP32-C6, ESP32-H2 (RISC-V), got '{}'",
-                        board.mcu
-                    )));
-                }
-                check_qemu_flash_mode(&board)?;
-                Ok(Box::new(QemuRunner::new(
-                    project_dir.to_path_buf(),
-                    env_name.to_string(),
-                    board,
-                )))
-            }
-            "avr8js" => {
-                if !matches!(
-                    platform,
-                    fbuild_core::Platform::AtmelAvr | fbuild_core::Platform::AtmelMegaAvr
-                ) {
-                    return Err(fbuild_core::FbuildError::DeployFailed(
-                        "avr8js runner is only supported for AVR boards".to_string(),
-                    ));
-                }
-                if !board.mcu.eq_ignore_ascii_case("atmega328p") {
-                    return Err(fbuild_core::FbuildError::DeployFailed(format!(
-                        "avr8js runner currently supports only ATmega328P, got '{}'",
-                        board.mcu
-                    )));
-                }
-                Ok(Box::new(Avr8jsRunner::new(board)))
-            }
-            "simavr" => {
-                if !matches!(
-                    platform,
-                    fbuild_core::Platform::AtmelAvr | fbuild_core::Platform::AtmelMegaAvr
-                ) {
-                    return Err(fbuild_core::FbuildError::DeployFailed(
-                        "simavr runner is only supported for AVR boards".to_string(),
-                    ));
-                }
-                if !board.has_emulator("simavr") {
-                    return Err(fbuild_core::FbuildError::DeployFailed(format!(
-                        "board '{}' does not advertise simavr support in its debug_tools",
-                        board_id
-                    )));
-                }
-                Ok(Box::new(SimavrRunner::new(board)))
-            }
-            other => Err(fbuild_core::FbuildError::DeployFailed(format!(
-                "unsupported emulator '{}'; available: qemu, avr8js, simavr",
-                other
-            ))),
-        };
-    }
-
-    // Auto-detect based on platform and MCU
-    match platform {
-        fbuild_core::Platform::AtmelAvr | fbuild_core::Platform::AtmelMegaAvr => {
-            if board.mcu.eq_ignore_ascii_case("atmega328p") {
-                // Default to avr8js for ATmega328P (no external binary needed)
-                Ok(Box::new(Avr8jsRunner::new(board)))
-            } else if board.has_emulator("simavr") {
-                // Use simavr for other AVR MCUs that advertise it
-                Ok(Box::new(SimavrRunner::new(board)))
-            } else {
-                Err(fbuild_core::FbuildError::DeployFailed(format!(
-                    "no emulator runner available for AVR MCU '{}'; \
-                     ATmega328P is supported via avr8js, other AVR boards require simavr in debug_tools",
-                    board.mcu
-                )))
-            }
-        }
-        fbuild_core::Platform::Espressif32 => {
-            if is_qemu_supported_esp32_mcu(&board.mcu) {
-                check_qemu_flash_mode(&board)?;
-                Ok(Box::new(QemuRunner::new(
-                    project_dir.to_path_buf(),
-                    env_name.to_string(),
-                    board,
-                )))
-            } else {
-                Err(fbuild_core::FbuildError::DeployFailed(format!(
-                    "no emulator runner available for ESP32 MCU '{}'; \
-                     ESP32, ESP32-S3 (Xtensa) and ESP32-C3, ESP32-C6, ESP32-H2 (RISC-V) \
-                     are supported via QEMU",
-                    board.mcu
-                )))
-            }
-        }
-        _ => Err(fbuild_core::FbuildError::DeployFailed(format!(
-            "no emulator runner available for platform {:?}",
-            platform
-        ))),
-    }
-}
-
-/// POST /api/test-emu handler — build firmware then run it in an emulator.
-pub async fn test_emu(
-    State(ctx): State<Arc<DaemonContext>>,
-    Json(req): Json<crate::models::TestEmuRequest>,
-) -> (StatusCode, Json<OperationResponse>) {
-    let request_id = req
-        .request_id
-        .clone()
-        .unwrap_or_else(|| uuid::Uuid::new_v4().to_string());
-    let project_dir = PathBuf::from(&req.project_dir);
-
-    // Mark the daemon as busy for the full build + emulate lifecycle.
-    // Without this guard the 30 s self-eviction loop sees an "empty"
-    // daemon during long (>30 s) ESP32/QEMU builds and triggers graceful
-    // shutdown, which closes the in-flight HTTP connection and surfaces
-    // as `error sending request for url (.../api/test-emu)` on the CLI
-    // side. See issue #130.
-    let _op_guard = crate::handlers::operations::OperationGuard::new(
-        &ctx,
-        fbuild_core::DaemonState::Building,
-        Some(format!("test-emu {}", req.project_dir)),
-    );
-
-    if !project_dir.exists() {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(OperationResponse::fail(
-                request_id,
-                format!("project directory does not exist: {}", req.project_dir),
-            )),
-        );
-    }
-
-    // Parse config
-    let config =
-        match fbuild_config::PlatformIOConfig::from_path(&project_dir.join("platformio.ini")) {
-            Ok(c) => c,
-            Err(e) => {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(OperationResponse::fail(
-                        request_id,
-                        format!("failed to parse platformio.ini: {}", e),
-                    )),
-                );
-            }
-        };
-
-    let env_name = req
-        .environment
-        .clone()
-        .or_else(|| config.get_default_environment().map(|s| s.to_string()))
-        .unwrap_or_else(|| "default".to_string());
-
-    let env_config = match config.get_env_config(&env_name) {
-        Ok(c) => c,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("invalid environment '{}': {}", env_name, e),
-                )),
-            );
-        }
-    };
-
-    let platform_str = env_config.get("platform").cloned().unwrap_or_default();
-    let platform = match fbuild_core::Platform::from_platform_str(&platform_str) {
-        Some(p) => p,
-        None => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("unsupported platform: {}", platform_str),
-                )),
-            );
-        }
-    };
-
-    let board_id = env_config.get("board").cloned().unwrap_or_else(|| {
-        fbuild_build::get_platform_support(platform)
-            .map(|s| s.default_board_id().to_string())
-            .unwrap_or_else(|_| "unknown".to_string())
-    });
-    let board_overrides = config.get_board_overrides(&env_name).unwrap_or_default();
-
-    // Select the emulator runner before building (fail fast on unsupported boards)
-    let runner = match select_runner(
-        &project_dir,
-        &env_name,
-        platform,
-        &board_id,
-        &board_overrides,
-        req.emulator.as_deref(),
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(OperationResponse::fail(request_id, e.to_string())),
-            );
-        }
-    };
-
-    // Build firmware (hold project lock only during build, not the emulator phase)
-    let build_result = {
-        let lock = ctx.project_lock(&project_dir);
-        let _guard = lock.lock().await;
-
-        let needs_qemu_flags = platform == fbuild_core::Platform::Espressif32
-            && req.emulator.as_deref() != Some("avr8js");
-        let board_for_flags = if needs_qemu_flags {
-            fbuild_config::BoardConfig::from_board_id(&board_id, &board_overrides).ok()
-        } else {
-            None
-        };
-
-        let build_dir = fbuild_paths::get_project_build_root(&project_dir);
-        let params = fbuild_build::BuildParams {
-            project_dir: project_dir.clone(),
-            env_name: env_name.clone(),
-            clean: false,
-            profile: fbuild_core::BuildProfile::Release,
-            build_dir,
-            verbose: req.verbose,
-            jobs: None,
-            generate_compiledb: false,
-            compiledb_only: false,
-            log_sender: None,
-            symbol_analysis: false,
-            symbol_analysis_path: None,
-            no_timestamp: false,
-            src_dir: None,
-            pio_env: req.pio_env.clone(),
-            extra_build_flags: if needs_qemu_flags {
-                board_for_flags
-                    .as_ref()
-                    .map(|b| crate::handlers::operations::qemu_extra_build_flags(platform, &b.mcu))
-                    .unwrap_or_default()
-            } else {
-                Vec::new()
-            },
-            watch_set_cache: Some(std::sync::Arc::clone(&ctx.watch_set_cache) as std::sync::Arc<_>),
-        };
-
-        let p = platform;
-        tokio::task::spawn_blocking(move || {
-            let orchestrator = fbuild_build::get_orchestrator(p)?;
-            orchestrator.build(&params)
-        })
-        .await
-    };
-
-    let (firmware_path, elf_path) = match build_result {
-        Ok(Ok(r)) if r.success => {
-            let fw = r.firmware_path.clone().unwrap_or_else(|| {
-                r.elf_path
-                    .clone()
-                    .unwrap_or_else(|| PathBuf::from("firmware.bin"))
-            });
-            (fw, r.elf_path)
-        }
-        Ok(Ok(r)) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("build failed: {}", r.message),
-                )),
-            );
-        }
-        Ok(Err(e)) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("build error: {}", e),
-                )),
-            );
-        }
-        Err(e) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("build task panicked: {}", e),
-                )),
-            );
-        }
-    };
-
-    // Run the emulator
-    let artifact_bundle = EmulatorArtifactBundle::from_paths(&firmware_path, elf_path.as_deref());
-    let run_config = EmulatorRunConfig {
-        firmware_path,
-        elf_path,
-        artifact_bundle,
-        timeout: req.timeout,
-        halt_on_error: req.halt_on_error.clone(),
-        halt_on_success: req.halt_on_success.clone(),
-        expect: req.expect.clone(),
-        show_timestamp: req.show_timestamp,
-        verbose: req.verbose,
-    };
-
-    let emu_result = match runner.run(&run_config).await {
-        Ok(r) => r,
-        Err(e) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(OperationResponse::fail(
-                    request_id,
-                    format!("emulator error: {}", e),
-                )),
-            );
-        }
-    };
-
-    let success = emu_result.is_success();
-    let exit_code = emu_result.exit_code.unwrap_or(if success { 0 } else { 1 });
-    let message = format!(
-        "{} test-emu {}: {}",
-        runner.name(),
-        if success { "passed" } else { "failed" },
-        emu_result.outcome
-    );
-
-    (
-        StatusCode::OK,
-        Json(OperationResponse {
-            success,
-            request_id,
-            message,
-            exit_code,
-            output_file: None,
-            output_dir: None,
-            launch_url: None,
-            stdout: Some(emu_result.stdout),
-            stderr: Some(emu_result.stderr),
-        }),
-    )
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use fbuild_build::{BuildOrchestrator, BuildParams};
-    use fbuild_core::BuildProfile;
-
-    #[test]
-    fn monitor_outcome_to_emulator_maps_success() {
-        let outcome = monitor_outcome_to_emulator(MonitorOutcome::Success("ok".into()), Some(0));
-        assert_eq!(outcome, EmulatorOutcome::Passed("ok".into()));
-    }
-
-    #[test]
-    fn monitor_outcome_to_emulator_maps_error() {
-        let outcome = monitor_outcome_to_emulator(MonitorOutcome::Error("bad".into()), Some(1));
-        assert_eq!(outcome, EmulatorOutcome::Failed("bad".into()));
-    }
-
-    #[test]
-    fn monitor_outcome_to_emulator_maps_crash() {
-        let outcome = monitor_outcome_to_emulator(
-            MonitorOutcome::Error("abort() was called at PC 0x4200".into()),
-            Some(134),
-        );
-        assert_eq!(
-            outcome,
-            EmulatorOutcome::Crashed("abort() was called at PC 0x4200".into())
-        );
-    }
-
-    #[test]
-    fn monitor_outcome_to_emulator_maps_timeout() {
-        let outcome =
-            monitor_outcome_to_emulator(MonitorOutcome::Timeout { expect_found: true }, None);
-        assert_eq!(outcome, EmulatorOutcome::TimedOut { expect_found: true });
-    }
-
-    fn test_process_command(lines: &[&str]) -> (PathBuf, Vec<String>) {
-        #[cfg(windows)]
-        {
-            let system_root =
-                std::env::var("SystemRoot").unwrap_or_else(|_| "C:\\Windows".to_string());
-            let exe = PathBuf::from(system_root).join(r"System32\cmd.exe");
-            let script = lines
-                .iter()
-                .map(|line| format!("echo {}", line))
-                .collect::<Vec<_>>()
-                .join(" & ");
-            (exe, vec!["/C".to_string(), script])
-        }
-
-        #[cfg(not(windows))]
-        {
-            let script = lines
-                .iter()
-                .map(|line| format!("printf '%s\\n' '{}'", line.replace('\'', "'\"'\"'")))
-                .collect::<Vec<_>>()
-                .join("; ");
-            (PathBuf::from("sh"), vec!["-c".to_string(), script])
-        }
-    }
-
-    #[tokio::test]
-    async fn run_qemu_process_reports_expected_success_output() {
-        let (exe, args) = test_process_command(&["Hello from ESP32-S3!"]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: None,
-                halt_on_success: None,
-                expect: Some("Hello from ESP32-S3"),
-                show_timestamp: false,
-                verbose: false,
-                process_label: "QEMU",
-            },
-        )
-        .await
-        .unwrap();
-
-        assert!(result.stdout.contains("Hello from ESP32-S3!"));
-        match result.outcome {
-            MonitorOutcome::Success(message) => {
-                assert!(message.contains("QEMU exited normally"));
-            }
-            other => panic!("expected success outcome, got {:?}", other),
-        }
-    }
-
-    #[tokio::test]
-    async fn run_qemu_process_surfaces_crash_decoder_output() {
-        let (exe, args) =
-            test_process_command(&["abort() was called at PC 0x42002a3c", "Rebooting..."]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: Some("no firmware\\.elf found"),
-                halt_on_success: None,
-                expect: None,
-                show_timestamp: false,
-                verbose: false,
-                process_label: "QEMU",
-            },
-        )
-        .await
-        .unwrap();
-
-        assert!(result
-            .stdout
-            .contains("abort() was called at PC 0x42002a3c"));
-        assert!(result.stdout.contains("no firmware.elf found"));
-        match result.outcome {
-            MonitorOutcome::Error(message) => {
-                assert!(message.contains("halt-on-error pattern matched"));
-            }
-            other => panic!("expected error outcome, got {:?}", other),
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn run_real_esp32s3_fixture_in_qemu() {
-        let repo_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .unwrap()
-            .parent()
-            .unwrap()
-            .to_path_buf();
-        let project_dir = repo_root.join("tests/platform/esp32s3");
-        if !project_dir.exists() {
-            eprintln!("SKIP: {} does not exist", project_dir.display());
-            return;
-        }
-
-        let build_dir = project_dir.join(".fbuild/build-qemu");
-        let params = BuildParams {
-            project_dir: project_dir.clone(),
-            env_name: "esp32s3".to_string(),
-            clean: true,
-            profile: BuildProfile::Release,
-            build_dir,
-            verbose: true,
-            jobs: None,
-            generate_compiledb: false,
-            compiledb_only: false,
-            log_sender: None,
-            symbol_analysis: false,
-            symbol_analysis_path: None,
-            no_timestamp: false,
-            src_dir: None,
-            pio_env: Default::default(),
-            extra_build_flags: vec![
-                "-DARDUINO_USB_MODE=0".to_string(),
-                "-DARDUINO_USB_CDC_ON_BOOT=0".to_string(),
-            ],
-            watch_set_cache: None,
-        };
-
-        let orchestrator = fbuild_build::esp32::orchestrator::Esp32Orchestrator;
-        let build_result = orchestrator
-            .build(&params)
-            .expect("ESP32-S3 fixture build should succeed");
-        assert!(build_result.success);
-
-        let firmware_path = build_result
-            .firmware_path
-            .clone()
-            .expect("should produce firmware.bin");
-        let elf_path = build_result.elf_path.clone();
-
-        let board =
-            fbuild_config::BoardConfig::from_board_id("esp32-s3-devkitc-1", &Default::default())
-                .unwrap();
-        let mcu_config = fbuild_build::esp32::mcu_config::get_mcu_config("esp32s3").unwrap();
-        let flash_size_bytes = fbuild_deploy::esp32::resolve_qemu_flash_size_bytes(
-            &board,
-            mcu_config.default_flash_size(),
-        )
-        .unwrap();
-
-        let session_dir = tempfile::TempDir::new().unwrap();
-        let flash_image = session_dir.path().join("flash.bin");
-        fbuild_deploy::esp32::create_qemu_flash_image(
-            &firmware_path,
-            &flash_image,
-            flash_size_bytes,
-            mcu_config.bootloader_offset(),
-            mcu_config.partitions_offset(),
-            mcu_config.firmware_offset(),
-            elf_path.as_deref(),
-        )
-        .unwrap();
-
-        let qemu = fbuild_packages::toolchain::EspQemuXtensa::new(&project_dir)
-            .and_then(|pkg| pkg.resolve_executable())
-            .expect("native QEMU should resolve for ignored integration test");
-        let args = fbuild_deploy::esp32::build_qemu_args(
-            &board.mcu,
-            &flash_image,
-            board.qemu_esp32_psram_config(),
-        );
-        let addr2line_path = resolve_esp32_toolchain_gcc_path(&project_dir, &mcu_config)
-            .ok()
-            .and_then(|gcc| fbuild_serial::crash_decoder::derive_addr2line_path(&gcc));
-
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        let result = rt
-            .block_on(run_qemu_process(
-                &qemu,
-                &args,
-                RunQemuOptions {
-                    elf_path,
-                    addr2line_path,
-                    timeout_secs: Some(15.0),
-                    halt_on_error: None,
-                    halt_on_success: Some("Hello from ESP32-S3!"),
-                    expect: Some("Hello from ESP32-S3!"),
-                    show_timestamp: false,
-                    verbose: true,
-                    process_label: "QEMU",
-                },
-            ))
-            .unwrap();
-
-        assert!(result.stdout.contains("Hello from ESP32-S3!"));
-        match result.outcome {
-            MonitorOutcome::Success(_) => {}
-            other => panic!("expected success outcome, got {:?}", other),
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // avr8js headless tests (use fake process, no real Node.js needed)
-    // -----------------------------------------------------------------------
-
-    #[tokio::test]
-    async fn run_avr8js_headless_captures_stdout() {
-        let (exe, args) = test_process_command(&["Hello from AVR!"]);
-        // run_avr8js_headless expects (node, script, hex, f_cpu, cache_dir, options).
-        // We bypass that by calling the lower-level run with the fake exe directly.
-        // Since run_avr8js_headless builds its own command, we test the same
-        // subprocess loop via run_qemu_process which shares identical logic.
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: None,
-                halt_on_success: None,
-                expect: Some("Hello from AVR"),
-                show_timestamp: false,
-                verbose: false,
-                process_label: "QEMU",
-            },
-        )
-        .await
-        .unwrap();
-
-        assert!(result.stdout.contains("Hello from AVR!"));
-        match result.outcome {
-            MonitorOutcome::Success(msg) => {
-                assert!(msg.contains("QEMU exited normally"));
-            }
-            other => panic!("expected success, got {:?}", other),
-        }
-    }
-
-    #[tokio::test]
-    async fn run_avr8js_headless_halt_on_success() {
-        let (exe, args) =
-            test_process_command(&["booting...", "PASS: all tests passed", "more output"]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: None,
-                halt_on_success: Some("PASS:"),
-                expect: None,
-                show_timestamp: false,
-                verbose: false,
-                process_label: "QEMU",
-            },
-        )
-        .await
-        .unwrap();
-
-        match result.outcome {
-            MonitorOutcome::Success(msg) => {
-                assert!(msg.contains("halt-on-success pattern matched"));
-            }
-            other => panic!("expected success, got {:?}", other),
-        }
-    }
-
-    #[tokio::test]
-    async fn run_avr8js_headless_halt_on_error() {
-        let (exe, args) = test_process_command(&["booting...", "FAIL: assertion failed"]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: Some("FAIL:"),
-                halt_on_success: None,
-                expect: None,
-                show_timestamp: false,
-                verbose: false,
-                process_label: "QEMU",
-            },
-        )
-        .await
-        .unwrap();
-
-        match result.outcome {
-            MonitorOutcome::Error(msg) => {
-                assert!(msg.contains("halt-on-error pattern matched"));
-            }
-            other => panic!("expected error, got {:?}", other),
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // SimAVR runner tests (use fake process, no real simavr needed)
-    // -----------------------------------------------------------------------
-
-    #[tokio::test]
-    async fn simavr_runner_captures_stdout_via_process_runner() {
-        // SimavrRunner delegates to run_qemu_process, so we verify the same
-        // subprocess monitoring path works for simavr-style output.
-        let (exe, args) = test_process_command(&["Hello from ATmega2560!"]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: None,
-                halt_on_success: None,
-                expect: Some("Hello from ATmega2560"),
-                show_timestamp: false,
-                verbose: false,
-                process_label: "simavr",
-            },
-        )
-        .await
-        .unwrap();
-
-        assert!(result.stdout.contains("Hello from ATmega2560!"));
-        match result.outcome {
-            MonitorOutcome::Success(_) => {}
-            other => panic!("expected success, got {:?}", other),
-        }
-    }
-
-    #[tokio::test]
-    async fn simavr_runner_halt_on_success() {
-        let (exe, args) = test_process_command(&["simavr: init", "PASS: all tests passed", "done"]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: None,
-                halt_on_success: Some("PASS:"),
-                expect: None,
-                show_timestamp: false,
-                verbose: false,
-                process_label: "simavr",
-            },
-        )
-        .await
-        .unwrap();
-
-        match result.outcome {
-            MonitorOutcome::Success(msg) => {
-                assert!(msg.contains("halt-on-success pattern matched"));
-            }
-            other => panic!("expected success, got {:?}", other),
-        }
-    }
-
-    #[tokio::test]
-    async fn simavr_runner_halt_on_error() {
-        let (exe, args) = test_process_command(&["simavr: init", "FAIL: assertion failed"]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(2.0),
-                halt_on_error: Some("FAIL:"),
-                halt_on_success: None,
-                expect: None,
-                show_timestamp: false,
-                verbose: false,
-                process_label: "simavr",
-            },
-        )
-        .await
-        .unwrap();
-
-        match result.outcome {
-            MonitorOutcome::Error(msg) => {
-                assert!(msg.contains("halt-on-error pattern matched"));
-            }
-            other => panic!("expected error, got {:?}", other),
-        }
-    }
-
-    #[tokio::test]
-    async fn simavr_runner_timeout() {
-        let (exe, args) = test_process_command(&["simavr: init", "running..."]);
-        let result = run_qemu_process(
-            &exe,
-            &args,
-            RunQemuOptions {
-                elf_path: None,
-                addr2line_path: None,
-                timeout_secs: Some(0.1),
-                halt_on_error: None,
-                halt_on_success: Some("PASS:"),
-                expect: None,
-                show_timestamp: false,
-                verbose: false,
-                process_label: "simavr",
-            },
-        )
-        .await
-        .unwrap();
-
-        // Process exits quickly so the outcome depends on whether halt pattern matched
-        // before the process ended — either timeout or a normal exit without the pattern.
-        match result.outcome {
-            MonitorOutcome::Success(_) => {
-                // Process exited cleanly before timeout, expect not set so counts as success
-            }
-            MonitorOutcome::Timeout { .. } => {
-                // Timed out waiting for halt-on-success pattern — also valid
-            }
-            MonitorOutcome::Error(_) => {
-                // Process exited before halt pattern found — also acceptable
-            }
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // select_runner tests for simavr
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn select_runner_explicit_simavr_for_uno() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "uno",
-            fbuild_core::Platform::AtmelAvr,
-            "uno",
-            &HashMap::new(),
-            Some("simavr"),
-        );
-        assert!(result.is_ok(), "select_runner should accept simavr for uno");
-        assert_eq!(result.unwrap().name(), "simavr");
-    }
-
-    #[test]
-    fn select_runner_explicit_simavr_for_mega() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "megaatmega2560",
-            fbuild_core::Platform::AtmelAvr,
-            "megaatmega2560",
-            &HashMap::new(),
-            Some("simavr"),
-        );
-        assert!(
-            result.is_ok(),
-            "select_runner should accept simavr for mega"
-        );
-        assert_eq!(result.unwrap().name(), "simavr");
-    }
-
-    #[test]
-    fn select_runner_explicit_simavr_for_leonardo() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "leonardo",
-            fbuild_core::Platform::AtmelAvr,
-            "leonardo",
-            &HashMap::new(),
-            Some("simavr"),
-        );
-        assert!(
-            result.is_ok(),
-            "select_runner should accept simavr for leonardo"
-        );
-        assert_eq!(result.unwrap().name(), "simavr");
-    }
-
-    #[test]
-    fn select_runner_explicit_simavr_rejects_esp32() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32dev",
-            fbuild_core::Platform::Espressif32,
-            "esp32dev",
-            &HashMap::new(),
-            Some("simavr"),
-        );
-        assert!(result.is_err(), "simavr should reject ESP32 boards");
-    }
-
-    #[test]
-    fn select_runner_auto_detects_simavr_for_mega() {
-        // ATmega2560 should auto-detect simavr since it's not ATmega328P
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "megaatmega2560",
-            fbuild_core::Platform::AtmelAvr,
-            "megaatmega2560",
-            &HashMap::new(),
-            None,
-        );
-        assert!(
-            result.is_ok(),
-            "auto-detect should find simavr for mega: {:?}",
-            result.err()
-        );
-        assert_eq!(result.unwrap().name(), "simavr");
-    }
-
-    #[test]
-    fn select_runner_auto_detects_simavr_for_leonardo() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "leonardo",
-            fbuild_core::Platform::AtmelAvr,
-            "leonardo",
-            &HashMap::new(),
-            None,
-        );
-        assert!(
-            result.is_ok(),
-            "auto-detect should find simavr for leonardo: {:?}",
-            result.err()
-        );
-        assert_eq!(result.unwrap().name(), "simavr");
-    }
-
-    #[test]
-    fn select_runner_auto_detects_avr8js_for_uno() {
-        // ATmega328P should still default to avr8js, not simavr
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "uno",
-            fbuild_core::Platform::AtmelAvr,
-            "uno",
-            &HashMap::new(),
-            None,
-        );
-        assert!(result.is_ok());
-        assert_eq!(
-            result.unwrap().name(),
-            "avr8js ATmega328P",
-            "ATmega328P should default to avr8js"
-        );
-    }
-
-    #[test]
-    fn select_runner_simavr_name_matches() {
-        let board =
-            fbuild_config::BoardConfig::from_board_id("megaatmega2560", &HashMap::new()).unwrap();
-        let runner = SimavrRunner::new(board);
-        assert_eq!(runner.name(), "simavr");
-    }
-
-    // -----------------------------------------------------------------------
-    // select_runner tests for ESP32 QEMU (Issue #25)
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn select_runner_explicit_qemu_for_esp32dev() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32dev",
-            fbuild_core::Platform::Espressif32,
-            "esp32dev",
-            &HashMap::new(),
-            Some("qemu"),
-        );
-        assert!(
-            result.is_ok(),
-            "select_runner should accept qemu for esp32dev: {:?}",
-            result.err()
-        );
-        assert_eq!(result.unwrap().name(), "QEMU ESP32");
-    }
-
-    #[test]
-    fn select_runner_explicit_qemu_for_esp32s3() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32-s3-devkitc-1",
-            fbuild_core::Platform::Espressif32,
-            "esp32-s3-devkitc-1",
-            &HashMap::new(),
-            Some("qemu"),
-        );
-        assert!(
-            result.is_ok(),
-            "select_runner should accept qemu for esp32-s3: {:?}",
-            result.err()
-        );
-        assert_eq!(result.unwrap().name(), "QEMU ESP32S3");
-    }
-
-    #[test]
-    fn select_runner_explicit_qemu_accepts_esp32c3() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32-c3-devkitm-1",
-            fbuild_core::Platform::Espressif32,
-            "esp32-c3-devkitm-1",
-            &HashMap::new(),
-            Some("qemu"),
-        );
-        assert!(
-            result.is_ok(),
-            "QEMU should accept ESP32-C3 via qemu-system-riscv32: {:?}",
-            result.err()
-        );
-    }
-
-    #[test]
-    fn select_runner_explicit_qemu_accepts_esp32c6() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32-c6-devkitc-1",
-            fbuild_core::Platform::Espressif32,
-            "esp32-c6-devkitc-1",
-            &HashMap::new(),
-            Some("qemu"),
-        );
-        assert!(
-            result.is_ok(),
-            "QEMU should accept ESP32-C6 via qemu-system-riscv32: {:?}",
-            result.err()
-        );
-    }
-
-    #[test]
-    fn select_runner_explicit_qemu_accepts_esp32h2() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32-h2-devkitm-1",
-            fbuild_core::Platform::Espressif32,
-            "esp32-h2-devkitm-1",
-            &HashMap::new(),
-            Some("qemu"),
-        );
-        assert!(
-            result.is_ok(),
-            "QEMU should accept ESP32-H2 via qemu-system-riscv32: {:?}",
-            result.err()
-        );
-    }
-
-    #[test]
-    fn select_runner_explicit_qemu_rejects_esp32s2() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32-s2-saola-1",
-            fbuild_core::Platform::Espressif32,
-            "esp32-s2-saola-1",
-            &HashMap::new(),
-            Some("qemu"),
-        );
-        assert!(
-            result.is_err(),
-            "QEMU should reject ESP32-S2 (not emulated upstream)"
-        );
-    }
-
-    #[test]
-    fn select_runner_auto_detects_qemu_for_esp32dev() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32dev",
-            fbuild_core::Platform::Espressif32,
-            "esp32dev",
-            &HashMap::new(),
-            None,
-        );
-        assert!(
-            result.is_ok(),
-            "auto-detect should find qemu for esp32dev: {:?}",
-            result.err()
-        );
-        assert_eq!(result.unwrap().name(), "QEMU ESP32");
-    }
-
-    #[test]
-    fn select_runner_auto_detects_qemu_for_esp32s3() {
-        let result = select_runner(
-            Path::new("/tmp/test"),
-            "esp32-s3-devkitc-1",
-            fbuild_core::Platform::Espressif32,
-            "esp32-s3-devkitc-1",
-            &HashMap::new(),
-            None,
-        );
-        assert!(
-            result.is_ok(),
-            "auto-detect should find qemu for esp32-s3: {:?}",
-            result.err()
-        );
-        assert_eq!(result.unwrap().name(), "QEMU ESP32S3");
-    }
-
-    #[test]
-    fn is_qemu_supported_esp32_mcu_accepts_xtensa() {
-        assert!(is_qemu_supported_esp32_mcu("esp32"));
-        assert!(is_qemu_supported_esp32_mcu("ESP32"));
-        assert!(is_qemu_supported_esp32_mcu("esp32s3"));
-        assert!(is_qemu_supported_esp32_mcu("ESP32S3"));
-    }
-
-    #[test]
-    fn is_qemu_supported_esp32_mcu_accepts_riscv() {
-        assert!(is_qemu_supported_esp32_mcu("esp32c3"));
-        assert!(is_qemu_supported_esp32_mcu("ESP32C3"));
-        assert!(is_qemu_supported_esp32_mcu("esp32c6"));
-        assert!(is_qemu_supported_esp32_mcu("esp32h2"));
-    }
-
-    #[test]
-    fn is_qemu_supported_esp32_mcu_rejects_unsupported() {
-        assert!(!is_qemu_supported_esp32_mcu("esp32s2"));
-        assert!(!is_qemu_supported_esp32_mcu("atmega328p"));
-        assert!(!is_qemu_supported_esp32_mcu("esp32p4"));
-    }
-
-    // -----------------------------------------------------------------------
-    // avr8js npm cache integrity tests (issue #86)
-    // -----------------------------------------------------------------------
-
-    /// Serialises tests that mutate process-wide env vars (PATH). Without
-    /// this, parallel cargo-test workers would clobber each other's PATH.
-    fn env_lock() -> std::sync::MutexGuard<'static, ()> {
-        use std::sync::{Mutex, OnceLock};
-        static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-        LOCK.get_or_init(|| Mutex::new(()))
-            .lock()
-            .unwrap_or_else(|p| p.into_inner())
-    }
-
-    #[test]
-    fn avr8js_cache_is_intact_detects_missing_dir() {
-        let tmp = tempfile::TempDir::new().unwrap();
-        // Empty cache dir: no node_modules at all.
-        assert!(!avr8js_cache_is_intact(tmp.path()));
-    }
-
-    #[test]
-    fn avr8js_cache_is_intact_detects_missing_package_json() {
-        let tmp = tempfile::TempDir::new().unwrap();
-        // Simulate a partial/corrupt install: dir exists, but no package.json.
-        std::fs::create_dir_all(tmp.path().join("node_modules").join("avr8js")).unwrap();
-        assert!(
-            !avr8js_cache_is_intact(tmp.path()),
-            "corrupt install (no package.json) must not count as intact"
-        );
-    }
-
-    #[test]
-    fn avr8js_cache_is_intact_accepts_marker_present() {
-        let tmp = tempfile::TempDir::new().unwrap();
-        let module_dir = tmp.path().join("node_modules").join("avr8js");
-        std::fs::create_dir_all(&module_dir).unwrap();
-        std::fs::write(module_dir.join("package.json"), b"{\"name\":\"avr8js\"}").unwrap();
-        assert!(avr8js_cache_is_intact(tmp.path()));
-    }
-
-    #[test]
-    fn prepare_avr8js_cache_skips_when_intact() {
-        let tmp = tempfile::TempDir::new().unwrap();
-        let module_dir = tmp.path().join("node_modules").join("avr8js");
-        std::fs::create_dir_all(&module_dir).unwrap();
-        std::fs::write(module_dir.join("package.json"), b"{}").unwrap();
-        assert_eq!(
-            prepare_avr8js_cache_for_install(tmp.path(), false),
-            Avr8jsCachePrep::AlreadyIntact
-        );
-        // Tree must survive untouched.
-        assert!(module_dir.join("package.json").exists());
-    }
-
-    #[test]
-    fn prepare_avr8js_cache_wipes_node_modules_when_corrupt() {
-        let tmp = tempfile::TempDir::new().unwrap();
-        let node_modules = tmp.path().join("node_modules");
-        let module_dir = node_modules.join("avr8js");
-        std::fs::create_dir_all(&module_dir).unwrap();
-        // Stray partial file inside an otherwise package.json-less install.
-        std::fs::write(module_dir.join("index.mjs"), b"partial").unwrap();
-        assert!(!avr8js_cache_is_intact(tmp.path()));
-
-        let prep = prepare_avr8js_cache_for_install(tmp.path(), false);
-        assert_eq!(prep, Avr8jsCachePrep::WipedNodeModules);
-        assert!(
-            !node_modules.exists(),
-            "corrupt node_modules tree must be wiped before reinstall"
-        );
-    }
-
-    #[test]
-    fn prepare_avr8js_cache_force_refresh_wipes_entire_dir() {
-        let tmp_parent = tempfile::TempDir::new().unwrap();
-        let cache = tmp_parent.path().join("avr8js-node");
-        let module_dir = cache.join("node_modules").join("avr8js");
-        std::fs::create_dir_all(&module_dir).unwrap();
-        std::fs::write(module_dir.join("package.json"), b"{}").unwrap();
-        // Even a fully-intact install must be wiped under force_refresh=true.
-        let prep = prepare_avr8js_cache_for_install(&cache, true);
-        assert_eq!(prep, Avr8jsCachePrep::ForceWiped);
-        assert!(!cache.exists(), "force-refresh must remove the cache dir");
-    }
-
-    #[test]
-    fn prepare_avr8js_cache_nothing_to_clean_on_fresh_path() {
-        let tmp_parent = tempfile::TempDir::new().unwrap();
-        let cache = tmp_parent.path().join("avr8js-node"); // doesn't exist yet
-        assert_eq!(
-            prepare_avr8js_cache_for_install(&cache, false),
-            Avr8jsCachePrep::NothingToClean
-        );
-    }
-
-    #[test]
-    fn refresh_emu_cache_requested_recognises_truthy_values() {
-        let _guard = env_lock();
-        for (val, expected) in [
-            ("1", true),
-            ("true", true),
-            ("TRUE", true),
-            ("yes", true),
-            ("YES", true),
-            ("0", false),
-            ("", false),
-            ("no", false),
-        ] {
-            std::env::set_var(REFRESH_EMU_CACHE_ENV, val);
-            assert_eq!(
-                refresh_emu_cache_requested(),
-                expected,
-                "{}={:?} should parse as {}",
-                REFRESH_EMU_CACHE_ENV,
-                val,
-                expected
-            );
-        }
-        std::env::remove_var(REFRESH_EMU_CACHE_ENV);
-        assert!(!refresh_emu_cache_requested());
-    }
-
-    /// When npm isn't on PATH, `ensure_avr8js_npm_in` must return an
-    /// `FbuildError::DeployFailed` that names both `npm` and the cache dir.
-    /// This is the fix for issue #86's silent `ERR_MODULE_NOT_FOUND`.
-    #[test]
-    fn ensure_avr8js_npm_in_reports_clear_error_without_npm() {
-        let _guard = env_lock();
-        let saved_path = std::env::var_os("PATH");
-        // PATHEXT matters on Windows for command resolution of .cmd files.
-        let saved_pathext = std::env::var_os("PATHEXT");
-
-        std::env::set_var("PATH", "");
-        #[cfg(windows)]
-        {
-            // Ensure .cmd isn't resolved via some fallback.
-            std::env::set_var("PATHEXT", "");
-        }
-
-        let tmp = tempfile::TempDir::new().unwrap();
-        let cache = tmp.path().join("avr8js-node");
-        let result = ensure_avr8js_npm_in(&cache, false);
-
-        // Restore BEFORE asserting so a panic doesn't leak PATH="" to sibling tests.
-        if let Some(p) = saved_path {
-            std::env::set_var("PATH", p);
-        } else {
-            std::env::remove_var("PATH");
-        }
-        if let Some(p) = saved_pathext {
-            std::env::set_var("PATHEXT", p);
-        } else {
-            std::env::remove_var("PATHEXT");
-        }
-
-        let err = result.expect_err("npm must be unresolvable with PATH=\"\"");
-        let msg = err.to_string();
-        assert!(
-            msg.contains("npm"),
-            "error message must mention 'npm'; got: {}",
-            msg
-        );
-        assert!(
-            msg.contains(&cache.display().to_string()),
-            "error message must include cache dir path; got: {}",
-            msg
-        );
-        assert!(
-            msg.contains(REFRESH_EMU_CACHE_ENV),
-            "error message must reference {} for recovery; got: {}",
-            REFRESH_EMU_CACHE_ENV,
-            msg
-        );
-    }
-
-    /// When the cache dir contains a corrupt partial install, the reinstall
-    /// path must fire (detected here by asserting the partial tree is wiped
-    /// even when the downstream npm call subsequently fails).
-    #[test]
-    fn ensure_avr8js_npm_in_wipes_corrupt_before_reinstall() {
-        let _guard = env_lock();
-        let saved_path = std::env::var_os("PATH");
-
-        // Force the npm spawn to fail so we isolate the wipe behaviour.
-        std::env::set_var("PATH", "");
-
-        let tmp = tempfile::TempDir::new().unwrap();
-        let cache = tmp.path().join("avr8js-node");
-        let node_modules = cache.join("node_modules");
-        let module_dir = node_modules.join("avr8js");
-        std::fs::create_dir_all(&module_dir).unwrap();
-        // Deliberately omit package.json → corrupt.
-        std::fs::write(module_dir.join("garbage"), b"partial").unwrap();
-        assert!(!avr8js_cache_is_intact(&cache));
-
-        let result = ensure_avr8js_npm_in(&cache, false);
-
-        if let Some(p) = saved_path {
-            std::env::set_var("PATH", p);
-        } else {
-            std::env::remove_var("PATH");
-        }
-
-        // npm spawn must fail (no PATH), but the corrupt tree must be gone.
-        assert!(result.is_err(), "empty PATH should prevent npm install");
-        assert!(
-            !node_modules.exists(),
-            "corrupt node_modules must be wiped before reinstall attempt"
-        );
-    }
-}
